% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol.R
\name{ALCDTTM}
\alias{ALCDTTM}
\title{Type of drinker (12 months)}
\usage{
ALCDTTM(ALCDTTM)
}
\arguments{
\item{ALCDTTM}{cchsflow variable name for type of drinker (12 months)}
}
\description{
\strong{NOTE:} this is not a function.

 This is a categorical variable derived by Statistics Canada that uses
 various intermediate alcohol variables to categorize individuals into 3
 distinct groups:

 \enumerate{
  \item Regular Drinker
  \item Occasional Drinker
  \item No drink in the last 12 months.
  }
}
\details{
This variable was introduced in the 2007-2008 cycle of the CCHS, and
 became the sole derived variable that categorized people into various
 drinker types from 2009 onwards. Unlike ALCDTYP, this variable does not
 distinguish between former and never drinkers.
}
\examples{
library(cchsflow)
 ?ALCDTTM

}
