% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol.R
\name{ALCDTYP_A}
\alias{ALCDTYP_A}
\title{ICES Type of drinker}
\usage{
ALCDTYP_A(ALCDTYP_A)
}
\arguments{
\item{ALCDTYP_A}{cchsflow variable name for type of drinker}
}
\description{
\strong{NOTE:} this is not a function.

 This is a categorical variable derived from ALCDTYP and ALCDTTM that uses
 various intermediate alcohol variables to categorize individuals into 2
 distinct groups:

 \enumerate{
  \item Former Drinker
  \item Other Drinker
 }

 For CCHS cycles from 2001 to 2007, categories regular drinker, 
 occasional drinker and never drinker are combined into an 
 "Other Drinker" category. Starting in 2007, Statistics Canada created 
 a derived variable that looked at drinking type in the last 12 months. 
 This new derived variable did not distinguish between former and never 
 drinkers. Therefore, for ICES CCHS cycles from 2007 onward regular drinker
 and occasional drinker are combined into the "Other Drinker" category
 and "Did not drink in the last 12 months" was categorized as "Former 
 Drinker". If your research requires you to differentiate between former 
 and never drinkers, we recommend using earlier cycles of the CCHS.
}
\examples{
library(cchsflow)
 ?ALCDTYP_A

}
