% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol-legacy.R
\name{ALW_1}
\alias{ALW_1}
\title{Any alcohol past week}
\usage{
ALW_1(ALW_1)
}
\arguments{
\item{ALW_1}{cchsflow variable name for any alcohol past week}
}
\description{
\strong{NOTE:} this is not a function.

 This is a categorical variable derived by Statistics Canada that determines 
 if alcohol was consumed in the past week. The variable is optional in 
 selected provinces and territories.
}
\details{
This variable is present in every CCHS cycle used in cchsflow. In 
2007 and 2008, the variable is optional for Newfoundland and Labrador, Nova 
Scotia, Ontario, British Columbia and Nunavut.In 2009 and 2010, the variable 
is optional for Newfoundland and Labrador, Ontario, and Saskatchewan. In 
2011, the variable is optional for Newfoundland and Labrador, Quebec, 
Ontario, Manitoba, and Saskatchewan. In 2012, the variable is optional for 
Newfoundland and Labrador, Quebec, Ontario, Manitoba, Nunavut, and 
Saskatchewan.In 2013, the variable is optional for Quebec, Ontario, Prince 
Edward Island, Manitoba, Yukon, and Saskatchewan. In 2014, the variable is 
optional for Nunavut, Quebec, Ontario, Prince Edward Island, Manitoba, 
Newfoundland and Labrador, Saskatchewan, and British Columbia.
}
\examples{
library(cchsflow)
 ?ALW_1

}
