% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol-legacy.R
\name{ALW_2A1}
\alias{ALW_2A1}
\title{Number of drinks - Sunday}
\usage{
ALW_2A1(ALW_2A1)
}
\arguments{
\item{ALW_2A1}{cchsflow variable name for number of drinks on Sunday}
}
\description{
\strong{NOTE:} this is not a function.

 This is a continuous variable derived by Statistics Canada that quantifies 
 the number of alcoholic drinks consumed on Sunday. The variable is 
 optional in selected provinces and territories.
}
\details{
This variable is present in every CCHS cycle used in cchsflow. In 
2007 and 2008, the variable is optional for Newfoundland and Labrador, Nova 
Scotia, Ontario, British Columbia and Nunavut.In 2009 and 2010, the variable 
is optional for Newfoundland and Labrador, Ontario, and Saskatchewan. In 
2011, the variable is optional for Newfoundland and Labrador, Quebec, 
Ontario, Manitoba, and Saskatchewan. In 2012, the variable is optional for 
Newfoundland and Labrador, Quebec, Ontario, Manitoba, Nunavut, and 
Saskatchewan.In 2013, the variable is optional for Quebec, Ontario, Prince 
Edward Island, Manitoba, Yukon, and Saskatchewan. In 2014, the variable is 
optional for Nunavut, Quebec, Ontario, Prince Edward Island, Manitoba, 
Newfoundland and Labrador, Saskatchewan, and British Columbia.
}
\examples{
library(cchsflow)
 ?ALW_2A1

}
