% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol-legacy.R
\name{ALWDDLY}
\alias{ALWDDLY}
\title{Average daily alcohol consumption}
\usage{
ALWDDLY(ALWDDLY)
}
\arguments{
\item{ALWDDLY}{cchsflow variable name for average daily alcohol consumption}
}
\description{
\strong{NOTE:} this is not a function.

 This is a continuous variable derived by Statistics Canada that quantifies
 the mean daily consumption of alcohol. This takes the value of ALWDWKY and
 divides it by 7.
}
\details{
This variable is present in every CCHS cycle used in cchsflow, and
 how it was derived remains consistent.
}
\examples{
library(cchsflow)
 ?ALWDDLY

}
