% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-variable-functions.R
\name{BMI_derived}
\alias{BMI_derived}
\title{Body Mass Index (BMI) derived variable}
\usage{
BMI_derived(HWTGHTM, HWTGWTK)
}
\arguments{
\item{HWTGHTM}{CCHS variable for height (in meters)}

\item{HWTGWTK}{CCHS variable for weight (in kilograms)}
}
\value{
a value for BMI
}
\description{
This function creates a derived BMI variable independent from the BMI variable provided by the CCHS.
}
\details{
This variable uses the CCHS variables for height and weight.
}
