% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking-status.R
\name{calculate_SMKDSTY_original}
\alias{calculate_SMKDSTY_original}
\title{Type of Smoker - SMKDSTY_original (deprecated alias)}
\usage{
calculate_SMKDSTY_original(...)
}
\arguments{
\item{...}{Arguments passed to [calculate_SMKDSTY_cat6()]}
}
\value{
Vector of smoking type classifications (1-6, plus missing value codes)
}
\description{
Deprecated alias for [calculate_SMKDSTY_cat6()]. Use
  `calculate_SMKDSTY_cat6()` directly for new code.
}
\examples{
\dontrun{
# Deprecated: use calculate_SMKDSTY_cat6() instead
calculate_SMKDSTY_original(SMK_005 = 1, SMK_030 = 2, SMK_01A = 1)
}

}
