% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{calculate_SMKG040}
\alias{calculate_SMKG040}
\title{Combine SMKG203_cont and SMKG207_cont into SMKG040}
\usage{
calculate_SMKG040(SMKG203_cont, SMKG207_cont)
}
\arguments{
\item{SMKG203_cont}{Continuous age started daily (current daily smokers)}

\item{SMKG207_cont}{Continuous age started daily (former daily smokers)}
}
\value{
Combined age started daily value
}
\description{
v3 alias for \code{\link{SMKG040_fun}}. Combines age-started-daily
  from daily smokers (SMKG203_cont) and former daily smokers (SMKG207_cont).
}
