% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent-time-canada.R
\name{categorize_pct_time}
\alias{categorize_pct_time}
\title{Categorize percent time in Canada}
\usage{
categorize_pct_time(pct_time_der)
}
\arguments{
\item{pct_time_der}{derived continuous percent time in Canada.
See \code{\link{calculate_pct_time}} for documentation on how variable was
derived.}
}
\value{
Character value for categorical percent time in Canada ("1" through
"10"), "NA(a)" for not applicable, or "NA(b)" for missing/invalid inputs.
}
\description{
Categorizes the derived percent time in Canada variable
(pct_time_der) into 10 percent intervals for pct_time_der_cat10.
}
\details{
The percent time in Canada provides an estimated percentage of the
time a person's life was spent in Canada. The categorical percent time in
Canada divides the continuous value into 10 percent intervals.

pct_time_der_cat10 uses the derived variable pct_time_der. pct_time_der uses
various variables that have been transformed by cchsflow (see documentation
on pct_time_der). In order to categorize percent time in Canada across CCHS
cycles, the variables must be transformed and harmonized.
}
\examples{
# Using rec_with_table() across CCHS cycles
library(cchsflow)
pct_time_cat2009_2010 <- rec_with_table(
  cchs2009_2010_p, c(
    "DHHGAGE_cont", "SDCGCBG",
    "SDCGRES_cont", "pct_time_der", "pct_time_der_cat10"
  )
)
head(pct_time_cat2009_2010)

pct_time_cat2011_2012 <- rec_with_table(
  cchs2011_2012_p, c(
    "DHHGAGE_cont", "SDCGCBG",
    "SDCGRES_cont", "pct_time_der", "pct_time_der_cat10"
  )
)
tail(pct_time_cat2011_2012)

combined_pct_time_cat <- merge_rec_data(pct_time_cat2009_2010,
pct_time_cat2011_2012)
head(combined_pct_time_cat)
tail(combined_pct_time_cat)

# Scalar usage
categorize_pct_time(55.5)

# Vector usage
categorize_pct_time(c(5, 25, 55, 85, 100))

}
