% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{categorize_quit_time}
\alias{categorize_quit_time}
\title{Time since quit categorization for former smokers}
\usage{
categorize_quit_time(time_value)
}
\arguments{
\item{time_value}{Continuous time since quitting value}
}
\value{
Categorized time with standard CCHS categories
}
\description{
Time since quit categorization for former smokers
}
