% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{categorize_time_ranges}
\alias{categorize_time_ranges}
\title{Time categorization helper for smoking cessation}
\usage{
categorize_time_ranges(time_value, breaks, labels)
}
\arguments{
\item{time_value}{Continuous time value}

\item{breaks}{Numeric vector of break points}

\item{labels}{Character vector of labels for categories}
}
\value{
Categorized time value
}
\description{
Time categorization helper for smoking cessation
}
\keyword{internal}
