% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{CCC_091_fun2}
\alias{CCC_091_fun2}
\title{CCC_091_fun2}
\usage{
CCC_091_fun2(CCC_91A, CCC_91E, CCC_91F)
}
\arguments{
\item{CCC_91A}{variable indicating if respondent has chronic bronchitis}

\item{CCC_91E}{variable indicating if respondent has emphysema}

\item{CCC_91F}{variable indicating if respondent has COPD}
}
\value{
a categorical variable (CCC_091) with 2 levels:

 \enumerate{
 \item respondent has COPD, emphysema, or chronic bronchitis
 \item respondent does not have COPD, emphysema, or chronic bronchitis
 }
}
\description{
This is one of 2 functions used to derive CCC_091
 (COPD/emphysema/chronic bronchitis) for CCHS cycles (2005-2008) in which
 the concept is split across three separate questions: CCC_91A (chronic
 bronchitis), CCC_91E (emphysema), and CCC_91F (COPD). A respondent is
 considered to have the condition if they answered yes to any question.
}
\examples{
# Using CCC_091_fun2() to derive CCC_091 for 2005/2007-08 cycles.
# CCC_091_fun2() is specified in variable_details.csv along with the
# CCHS variables and cycles included.

library(cchsflow)

CCC_091_2005 <- suppressWarnings(rec_with_table(
  cchs2005_p, c("CCC_91A", "CCC_91E", "CCC_91F", "CCC_091")
))

head(CCC_091_2005)
}
\seealso{
\code{\link{CCC_091_fun1}}
}
