% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2001}
\alias{cchs2001}
\title{2001 CCHS data}
\source{
http://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=3359
}
\value{
\item{cchs2001}{a data frame}
}
\description{
This is a subset of 200 observations from the 2001 cycle of the Canadian Community Health Survey (CCHS) 
Public Use Microdata file (PUMF) dataset.This survey is conducted by Statistics Canada. The CCHS 
has a [Statistics Canada Open License](https://www.statcan.gc.ca/eng/reference/licence).
}
\details{
Long name: cchs-82M0013-E-2001-c1-1-general-file

DDI: '~/inst/extdata/CCHS_DDI/CCHS_2001_DDI.xml
}
\examples{
data(cchs2001)
str(cchs2001)
}
\keyword{datasets}
