% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2005}
\alias{cchs2005}
\title{2005 CCHS data}
\source{
http://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=22642
}
\value{
\item{cchs2005}{a data frame}
}
\description{
This is a subset of 200 observations from the 2005 cycle of the CCHS PUMF dataset.
}
\details{
Long name: cchs-82M0013-E-2005-c3-1-main-file

DDI: '~/inst/extdata/CCHS_DDI/CCHS_2005_DDI.xml
}
\examples{
data(cchs2005)
str(cchs2005)
}
\keyword{datasets}
