% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2009_2010_p}
\alias{cchs2009_2010_p}
\title{2009-2010 CCHS PUMF subset data (200 respondents)}
\source{
\url{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=67251}
}
\value{
\item{cchs2009_2010_p}{a data frame}
}
\description{
This is a subset of 200 observations from the 2009-2010 cycle of the Canadian
Community Health Survey (CCHS) Public Use Microdata file (PUMF) dataset.
The CCHS survey is conducted by Statistics Canada.
}
\details{
See \href{https://www.statcan.gc.ca/en/reference/licence}{here} for the open
license. Source from Statistics Canada, Canadian Community Health Survey
PUMF, accessed Jan 2020. Reproduced and distributed on an "as is" basis with
the permission of Statistics Canada.

Long name: CCHS-82M0013-E-2009-2010-Annualcomponent

DDI: \url{https://osf.io/ynzpe/}

Additional documentation (PDFs): \url{https://osf.io/hkuy3/}
}
\examples{
data(cchs2009_2010_p)
str(cchs2009_2010_p)
}
\keyword{datasets}
