% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2013_2014}
\alias{cchs2013_2014}
\title{2013-2014 CCHS data}
\source{
http://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=144170
}
\value{
\item{cchs2013_2014}{a data frame}
}
\description{
This is a subset of 200 observations from the 2013-2014 cycle of the CCHS PUMF dataset.
}
\details{
Long name: cchs-82M0013-E-2013-2014-Annual-component

DDI: '~/inst/extdata/CCHS_DDI/CCHS_2013_2014_DDI.xml
}
\examples{
data(cchs2013_2014)
str(cchs2013_2014)
}
\keyword{datasets}
