% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2013_2014_p}
\alias{cchs2013_2014_p}
\title{2013-2014 CCHS PUMF subset data (200 respondents)}
\source{
\url{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=144170}
}
\value{
\item{cchs2013_2014_p}{a data frame}
}
\description{
This is a subset of 200 observations from the 2013-2014 cycle of the Canadian
Community Health Survey (CCHS) Public Use Microdata file (PUMF) dataset.
The CCHS survey is conducted by Statistics Canada.
}
\details{
See \href{https://www.statcan.gc.ca/en/reference/licence}{here} for the open
license. Source from Statistics Canada, Canadian Community Health Survey
PUMF, accessed Jan 2020. Reproduced and distributed on an "as is" basis with
the permission of Statistics Canada.

Long name: cchs-82M0013-E-2013-2014-Annual-component

DDI: \url{https://osf.io/gy25d/}

Additional documentation (PDFs): \url{https://osf.io/hkuy3/}
}
\examples{
data(cchs2013_2014_p)
str(cchs2013_2014_p)
}
\keyword{datasets}
