% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2014}
\alias{cchs2014}
\title{2014 CCHS data}
\source{
\url{http://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=164081}
}
\value{
\item{cchs2014}{a data frame}
}
\description{
This is a subset of 200 observations from the 2014 cycle of the Canadian
Community Health Survey (CCHS) Public Use Microdata file (PUMF) dataset.
The CCHS survey is conducted by Statistics Canada.
}
\details{
See \href{https://www.statcan.gc.ca/eng/reference/licence}{here} for the open
license.

Long name: cchs-82M0013-E-2014-Annual-component

DDI: \url{https://osf.io/tbmdn/}

Additional documentation (PDFs): \url{https://osf.io/hkuy3/}
}
\examples{
data(cchs2014)
str(cchs2014)
}
\keyword{datasets}
