% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{check_worksheet}
\alias{check_worksheet}
\title{Check a CSV worksheet for formatting errors}
\usage{
check_worksheet(file_path, file_type = c("variables", "variable_details"))
}
\arguments{
\item{file_path}{Path to the CSV file to check}

\item{file_type}{Type of file being checked. Either "variables" or
"variable_details".}
}
\value{
A list of errors found. Each error is a named list containing
information about the error.
}
\description{
Check a CSV worksheet for formatting errors
}
\examples{
\dontrun{
check_worksheet("inst/extdata/variables.csv", "variables")
}
}
