% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{classify_smoking_status}
\alias{classify_smoking_status}
\title{Smoking status classification helper}
\usage{
classify_smoking_status(smkdsty_value)
}
\arguments{
\item{smkdsty_value}{SMKDSTY smoking status value}
}
\value{
Character classification of smoking status
}
\description{
Smoking status classification helper
}
