% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-with-table.R
\name{compare_value_based_on_interval}
\alias{compare_value_based_on_interval}
\title{Compare Value Based On Interval}
\usage{
compare_value_based_on_interval(
  left_boundary,
  right_boundary,
  data_source,
  compare_columns,
  interval
)
}
\arguments{
\item{left_boundary}{the min value}

\item{right_boundary}{the max value}

\item{data_source}{the data that contains values being compared}

\item{compare_columns}{The columns inside data_source being checked}

\item{interval}{The scientific notation interval}
}
\value{
a boolean vector containing true for rows where the
comparison is true
}
\description{
Compare values on the scientific notation interval
}
