% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{derive_CCC_091_2001to2003}
\alias{derive_CCC_091_2001to2003}
\title{CCC_091_fun1}
\usage{
derive_CCC_091_2001to2003(CCC_91A, CCC_91B, output_format = "tagged_na")
}
\arguments{
\item{CCC_91A}{variable indicating if respondent has chronic bronchitis}

\item{CCC_91B}{variable indicating if respondent has emphysema or COPD}
}
\value{
a categorical variable (CCC_091) with 2 levels:

 \enumerate{
 \item respondent has COPD, emphysema, or chronic bronchitis
 \item respondent does not have COPD, emphysema, or chronic bronchitis
 }
}
\description{
This is one of 2 functions used to derive CCC_091
 (COPD/emphysema/chronic bronchitis) for CCHS cycles (2001-2003) in which
 the concept is split across two separate questions: CCC_91A (chronic
 bronchitis) and CCC_91B (emphysema/COPD combined). A respondent is
 considered to have the condition if they answered yes to either question.
}
\examples{
# Using CCC_091_fun1() to derive CCC_091 for 2001/2003 cycles.
# CCC_091_fun1() is specified in variable_details.csv along with the
# CCHS variables and cycles included.

library(cchsflow)

CCC_091_2001 <- suppressWarnings(rec_with_table(
  cchs2001_p, c("CCC_91A", "CCC_91B", "CCC_091")
))

head(CCC_091_2001)
}
\seealso{
\code{\link{CCC_091_fun2}}
}
