% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{derive_current_smoker}
\alias{derive_current_smoker}
\title{Current vs former smoker classification}
\usage{
derive_current_smoker(smkdsty_cat5)
}
\arguments{
\item{smkdsty_cat5}{5-category smoking status variable}
}
\value{
Binary current smoker indicator (1 = current, 0 = not current)
}
\description{
Current vs former smoker classification
}
\keyword{internal}
