% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{derive_ever_smoker}
\alias{derive_ever_smoker}
\title{Ever smoker classification}
\usage{
derive_ever_smoker(smkdsty_cat5)
}
\arguments{
\item{smkdsty_cat5}{5-category smoking status variable}
}
\value{
Binary ever smoker indicator (1 = ever smoked, 0 = never)
}
\description{
Ever smoker classification
}
