% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.append_to_list}
\alias{.append_to_list}
\title{Add an item to a list}
\usage{
.append_to_list(x, item)
}
\arguments{
\item{x}{the list}

\item{item}{the item to add}
}
\value{
the new list
}
\description{
Add an item to a list
}
