% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.check_row_sorting}
\alias{.check_row_sorting}
\title{Check the rows order in a worksheet}
\usage{
.check_row_sorting(csv_data, id_column_name, error_ctx)
}
\arguments{
\item{csv_data}{A data.frame containing the worksheet rows}

\item{id_column_name}{Name of the column to check for sorting}

\item{error_ctx}{Information used when creating the error object. A named
list with the following fields:
* file_type: The type of worksheet the CSV contains. Can be "variables" or
  "variable_details".
* file_path: The file path to the worksheet}
}
\value{
The list of row order errors found in the worksheet
}
\description{
Check the rows order in a worksheet
}
