% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.check_trailing_empty_columns}
\alias{.check_trailing_empty_columns}
\title{Check a worksheet for trailing empty columns}
\usage{
.check_trailing_empty_columns(csv_data, error_ctx)
}
\arguments{
\item{csv_data}{Data.frame containing the worksheet rows}

\item{error_ctx}{Information used when creating the error object. A named
list with the following fields:
* file_type: The type of worksheet the CSV contains. Can be "variables" or
  "variable_details".
* file_path: The file path to the worksheet}
}
\value{
List of errors
}
\description{
Check a worksheet for trailing empty columns
}
