% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-worksheet.R
\name{.fix_column_order_errors}
\alias{.fix_column_order_errors}
\title{Reorder columns to match expected order in a CSV worksheet}
\usage{
.fix_column_order_errors(csv_data, column_order_errors)
}
\arguments{
\item{csv_data}{A data frame containing the CSV data.}

\item{column_order_errors}{A list of column order error objects from
\code{check_worksheet}.}
}
\value{
The reordered data frame
}
\description{
Reorder columns to match expected order in a CSV worksheet
}
\keyword{internal}
