% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-worksheet.R
\name{.fix_empty_column_errors}
\alias{.fix_empty_column_errors}
\title{Remove empty trailing columns from CSV data}
\usage{
.fix_empty_column_errors(csv_data, empty_column_errors)
}
\arguments{
\item{csv_data}{A data frame containing the CSV data.}

\item{empty_column_errors}{A list of empty column error objects from
\code{check_worksheet}.}
}
\value{
The data frame with empty columns removed.
}
\description{
Remove empty trailing columns from CSV data
}
\keyword{internal}
