% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-worksheet.R
\name{.fix_extra_column_errors}
\alias{.fix_extra_column_errors}
\title{Remove extra columns not defined in the schema}
\usage{
.fix_extra_column_errors(csv_data, extra_column_errors)
}
\arguments{
\item{csv_data}{A data frame containing the CSV data.}

\item{extra_column_errors}{A list of extra column error objects from
\code{check_worksheet}.}
}
\value{
The data frame with extra columns removed.
}
\description{
Remove extra columns not defined in the schema
}
\keyword{internal}
