% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-worksheet.R
\name{.fix_unsorted_rows_error}
\alias{.fix_unsorted_rows_error}
\title{Sort rows alphabetically by the ID column}
\usage{
.fix_unsorted_rows_error(csv_data, unsorted_row_errors)
}
\arguments{
\item{csv_data}{A data frame containing the CSV data.}

\item{unsorted_row_errors}{A list of unsorted row error objects from
\code{check_worksheet}. Each error contains an \code{id_column_name} field.}
}
\value{
The sorted data frame
}
\description{
Sort rows alphabetically by the ID column
}
\keyword{internal}
