% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-discovery.R
\name{.load_cep_variables}
\alias{.load_cep_variables}
\title{Load variables from CEP worksheets}
\usage{
.load_cep_variables(cep_dir)
}
\description{
Load variables from CEP worksheets
}
\keyword{internal}
