% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-tools.R
\name{.load_variable_details}
\alias{.load_variable_details}
\title{Load variable_details.csv from package or repo}
\usage{
.load_variable_details(variable_details = NULL)
}
\description{
Load variable_details.csv from package or repo
}
\keyword{internal}
