% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-tools.R
\name{.load_variables_sheet}
\alias{.load_variables_sheet}
\title{Load variables.csv from package or repo}
\usage{
.load_variables_sheet(variables_sheet = NULL)
}
\description{
Load variables.csv from package or repo
}
\keyword{internal}
