% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.parse_csv_text}
\alias{.parse_csv_text}
\title{Parse the worksheet contents}
\usage{
.parse_csv_text(csv_text)
}
\arguments{
\item{csv_text}{String containing the contents of the worksheet}
}
\value{
A list of list of strings. The top level list contains the rows in
the worksheet. The second level contains the cell values of each row. For
example, for the following CSV file `a,b,c\n1,2,3` the result of this
function would be, `list(list('a', 'b', 'c'), list('1', '2', '3'))`
}
\description{
Parse the worksheet contents
}
