% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-discovery.R
\name{.parse_variable_start}
\alias{.parse_variable_start}
\title{Parse variableStart field into named list}
\usage{
.parse_variable_start(vs)
}
\description{
Parse variableStart field into named list
}
\keyword{internal}
