% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-tools.R
\name{.resolve_one}
\alias{.resolve_one}
\title{Recursively resolve dependencies for one variable}
\usage{
.resolve_one(
  variable,
  adj,
  variable_details,
  databases,
  graph,
  warnings,
  visited,
  stack
)
}
\description{
Recursively resolve dependencies for one variable
}
\keyword{internal}
