% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression.R
\name{DPSDPP}
\alias{DPSDPP}
\title{Depression Scale - Predicted Probability}
\usage{
DPSDPP(DPSDPP)
}
\arguments{
\item{DPSDPP}{cchsflow variable name for derived depression scale predicted
probability.}
}
\description{
\strong{NOTE:} this is not a function.

 This is categorical variable derived by Statistics Canada that predicts
 the probability that a respondent would be diagnosed as having a major
 depressive episode if a diagnostic interview was completed. This variable
 is derived from \code{\link{DPSDSF}} in which probabilities are assigned
 to respondents based on their depression scale score. For more details on
 how the variable was derived click \href{https://osf.io/auqwh/}{here}.
}
\details{
While this variable was considered to be categorical in CCHS
 documentation, the values range from 0 to 0.90 with no distinct names or
 metadata for each category. As such, this variable was specified as a
 continuous variable in cchsflow. This has no bearing on the final output
 of the variable as there are no recode changes. This means that a respondent
 who was coded with a probability of 0.50 will still have a probability value
 of 0.50 when the variable goes through harmonization.

 This variable is present in every CCHS cycle used in cchsflow, and
 how it was derived remains consistent.
}
\examples{
library(cchsflow)
?DPSDPP

}
\seealso{
\code{\link{DPSDSF}}
}
