% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-worksheet.R
\name{fix_worksheet}
\alias{fix_worksheet}
\title{Fix formatting issues in a CSV worksheet}
\usage{
fix_worksheet(file_path, file_type = c("variables", "variable_details"))
}
\arguments{
\item{file_path}{Path to the CSV file to fix}

\item{file_type}{Type of file being fixed. Either "variables" or
"variable_details".}
}
\value{
A list of error objects, each with an added \code{fixed} field:
  \itemize{
    \item fixed: Boolean indicating if the error was fixed
  }
  Returns an empty list if no errors were found.
}
\description{
Automatically corrects formatting violations in a CSV worksheet
  based on the errors detected by \code{check_worksheet}.
}
\examples{
\dontrun{
errors <- fix_worksheet("inst/extdata/variables.csv", "variables")
fixed_count <- sum(sapply(errors, function(e) e$fixed))
print(sprintf("Fixed \%d errors", fixed_count))
}
}
