% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-with-table.R
\name{get_data_variable_name}
\alias{get_data_variable_name}
\title{Get Data Variable Name}
\usage{
get_data_variable_name(
  data_name,
  data,
  row_being_checked,
  variable_being_checked
)
}
\arguments{
\item{data_name}{name of the database being checked}

\item{data}{database being checked}

\item{row_being_checked}{the row from variable details that contains
information on this variables}

\item{variable_being_checked}{the name of the recoded variable}
}
\value{
the data_source equivalent of variable_being_checked
}
\description{
Retrieves the name of the column inside data_source to
use for calculations
}
