% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worksheet-getters.R
\name{get_variable_type}
\alias{get_variable_type}
\title{Get Variable Type from Metadata}
\usage{
get_variable_type(var_name, variable_details = NULL)
}
\arguments{
\item{var_name}{Character. Variable name to check}

\item{variable_details}{Data.frame. Optional pre-loaded variable_details data}
}
\value{
Character. "continuous" or "categorical"
}
\description{
Determines if a variable is continuous or categorical from variable_details metadata.
}
