% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-variable-functions.R
\name{ifelse2}
\alias{ifelse2}
\title{Custom ifelse function that evaluates missing (NA) values}
\usage{
ifelse2(x, a, b)
}
\arguments{
\item{x}{A logical argument}

\item{a}{value if `x` is `TRUE`}

\item{b}{value if `x` is `FALSE`}
}
\value{
None
}
\description{
If the logical argument compares to a value that is NA, it will set `x` to `FALSE`
}
\examples{
ifelse2((age<18),"child", "adult")

}
