% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-with-table.R
\name{is_equal}
\alias{is_equal}
\title{is equal}
\usage{
is_equal(v1, v2)
}
\arguments{
\item{v1}{variable 1}

\item{v2}{variable 2}
}
\value{
boolean value of whether or not v1 and v2 are equal
}
\description{
Function to compare even with NA present
This function returns TRUE wherever elements are the same, including NA's,
and false everywhere else.
}
\examples{
library(cchsflow)
is_equal(1, 2)
# FALSE

is_equal(1, 1)
# TRUE

1 == NA
# NA

is_equal(1, NA)
# FALSE

NA == NA
# NA

is_equal(NA, NA)
# TRUE
}
