% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-generators.R
\name{is_primary_recommendation}
\alias{is_primary_recommendation}
\title{Check if variable is a primary recommendation}
\usage{
is_primary_recommendation(notes)
}
\arguments{
\item{notes}{Character. The notes field content}
}
\value{
Logical. TRUE if {recommended:primary} tag is present.
}
\description{
Convenience function to check for {recommended:primary} tag.
}
