% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-utils.R
\name{label_data}
\alias{label_data}
\title{label_data}
\usage{
label_data(label_list, data_to_label)
}
\arguments{
\item{label_list}{The label list that contains labels
from variable details}

\item{data_to_label}{The data that is to be labeled}
}
\value{
Returns labeled data
}
\description{
Labels a dataframe according to a given label list.
}
