% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupation.R
\name{LBFA_31A}
\alias{LBFA_31A}
\title{Occupation Group (9 categories)}
\usage{
LBFA_31A(LBFA_31A)
}
\arguments{
\item{LBFA_31A}{cchsflow variable name for Occupation Group (9 categories)}
}
\description{
\strong{NOTE:} this is not a function.

 This is a 9 category variable (LBFA_31A) that is in the CCHS that asks
 which occupation group best describes a respondent. Occupation group is
 asked in the 2001 CCHS cycle and in CCHS cycles from 2007-2014.
}
\details{
While occupation group is asked in many survey cycles, the 2001
 CCHS survey cycle is the only survey that has 9 categories. The categories
 are as follows:
 \enumerate{
 \item Management
 \item Professional (including accountants)
 \item Technologist, Technician or Tech Occupation
 \item Administrative, Financial or Clerical
 \item Sales or Service
 \item Trades, Transport or Equipment Operator
 \item Farming, Forestry, Fishing, Mining
 \item Processing, Manufacturing, Utilities
 \item Other
 }
 To harmonize the 2001 CCHS cycle with other survey cycles,
 \code{\link{LBFA_31A_a}} and \code{\link{LBFA_31A_b}} were created in which
 categories in the 2001 survey cycle were collapsed.
}
\examples{
library(cchsflow)
?LBFA_31A

}
\seealso{
\code{\link{LBFA_31A_a}}, \code{\link{LBFA_31A_b}}
}
