% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupation.R
\name{LBFA_31A_b}
\alias{LBFA_31A_b}
\title{Occupation Group (6 categories)}
\usage{
LBFA_31A_b(LBFA_31A_b)
}
\arguments{
\item{LBFA_31A_b}{cchsflow variable name for Occupation Group (6 categories)}
}
\description{
\strong{NOTE:} this is not a function.

 This is a 6 category variable (LBFA_31A_b) that is in the CCHS that asks
 which occupation group best describes a respondent. Occupation group is
 asked in the 2001 CCHS cycle and in CCHS cycles from 2007-2014.
}
\details{
In the 2007-2014 CCHS survey cycles, occupation group has 5
 categories. This variable, however, includes a sixth category to account for
 the "other" category asked in the 2001 CCHS survey cycle. The categories are
 as follows:
 \enumerate{
 \item Management, Health, Education, Art, Culture
 \item Business, Finance, Admin
 \item Sales or Service
 \item Trades, Transport or Equipment Operator
 \item Unique to Primary Industry/Processing/Manufacturing
 \item Other
 }

 In this variable, categories from the 2001 CCHS survey cycle were collapsed
 to harmonize with the other survey cycles. "Management, Professional
 (including accountants), Technologist, Technician or Tech Occupation" were
 combined into one category "Management, Health, Education, Art, Culture".
 "Farming, Forestry, Fishing, Mining" and "Processing, Manufacturing,
 Utilities", were combined into one category "Farming, Forestry, Fishing,
 Mining, Processing, Manufacturing, Utilities".
}
\examples{
library(cchsflow)
?LBFA_31A_b

}
\seealso{
\code{\link{LBFA_31A}}, \code{\link{LBFA_31A_a}}
}
