% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-discovery.R
\name{list_subjects}
\alias{list_subjects}
\title{List available subjects in variables.csv}
\usage{
list_subjects(source = "main")
}
\arguments{
\item{source}{Character. "main" or "cep"}
}
\value{
Character vector of unique subjects
}
\description{
List available subjects in variables.csv
}
