% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol-legacy.R
\name{low_drink_score_fun}
\alias{low_drink_score_fun}
\title{Low drinking score (all cycles)}
\usage{
low_drink_score_fun(DHH_SEX, ALWDWKY)
}
\arguments{
\item{DHH_SEX}{Sex of respondent (1 - male, 2 - female)}

\item{ALWDWKY}{Number of drinks consumed in the past week}
}
\value{
Low risk drinking score (low_drink_score) with four categories:
\itemize{
  \item 1 - Low risk (0 points)
  \item 2 - Marginal risk (1-2 points)
  \item 3 - Medium risk (3-4 points)
  \item 4 - High risk (5-9 points)
 }
}
\description{
This function creates a derived variable based on their drinking 
habits and flags for health and social problems from their pattern of 
alcohol use according to Canada's Low-Risk Alcohol Drinking Guideline.
}
\details{
The low risk drinking score is based on the scoring system in 
Canada's Low-Risk Alcohol Drinking Guideline. The score is divided into two 
steps. Step 1 allocates points based on sex and the number of drinks 
that you usually have each week. In step 2, one point will be awarded for 
each item that is true related to drinking habits. The total score is 
obtained from adding the points in step 1 and step 2.
}
\note{
Step 2 is not included in this function because the questions in 
step 2 are not asked in any of the CCHS cycles. The score is only based on 
step 1.

See \url{https://osf.io/eprg7/} for more details on the guideline and score.

v3.0.0, last updated: 2025-06-30, status: active, Note: Improved scoring algorithm with enhanced validation
}
\examples{
 
# Using low_drink_score_fun() to create low_drink_score values across 
# CCHS cycles low_drink_score_fun() is specified in variable_details.csv 
# along with the CCHS variables and cycles included.

# To transform low_drink_score, use rec_with_table() for each CCHS cycle
# and specify low_drink_score, along with the various alcohol and sex
# variables. 
# Using merge_rec_data(), you can combine low_drink_score across cycles.

library(cchsflow)
low_drink2001 <- rec_with_table(
  cchs2001_p, c(
    "DHH_SEX", "ALWDWKY", "low_risk_score"
  )
)

head(low_drink2001)

low_drink2009_2010 <- rec_with_table(
  cchs2009_2010_p, c(
    "DHH_SEX", "ALWDWKY", "low_risk_score"
  )
)

tail(low_drink2009_2010)

combined_low_drink <- bind_rows(low_drink2001, 
low_drink2009_2010)

head(combined_low_drink)

tail(combined_low_drink)

library(cchsflow)
DHH_SEX <- c(1, 2) 
ALWDWKY <- c(30, 15)

low_drink_score_data <- data.frame(DHH_SEX, ALWDWKY)
low_drink_score_data[,"Low_Drink_Score"] <- NA
print(low_drink_score_data)

low_drink_score_data$Low_Drink_Score <-low_drink_score_fun(low_drink_score_data[[1]], 
                                                       low_drink_score_data[[2]])
print(low_drink_score_data)
}
