% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-conditions.R
\name{multiple_conditions_fun}
\alias{multiple_conditions_fun}
\title{Multiple chronic conditions}
\usage{
multiple_conditions_fun(
  CCC_121,
  CCC_131,
  CCC_151,
  CCC_171,
  CCC_280,
  resp_condition,
  CCC_051
)
}
\arguments{
\item{CCC_121}{variable indicating if respondent has heart disease}

\item{CCC_131}{variable indicating if respondent has active cancer}

\item{CCC_151}{variable indicating if respondent suffers from the effects
of a stroke}

\item{CCC_171}{variable indicating if respondent has a bowel disorder}

\item{CCC_280}{variable indicating if respondent has a mood disorder. Note,
variable was not asked to respondents in the 2001 CCHS survey cycle}

\item{resp_condition}{derived variable indicating if respondent has a
respiratory condition. See \code{\link{resp_condition_fun1}} for
documentation on how variable was derived}

\item{CCC_051}{variable indicating if respondent has arthritis or
rheumatism}
}
\value{
A categorical variable indicating the number of chronic conditions
 a respondent has. Respondents with 5 or more conditions are grouped in the
 "5+" category.
}
\description{
This function generates a derived variable (multiple_conditions)
 that counts the number of chronic conditions a respondent has. This function
 takes 6 CCHS variables (heart disease, cancer, stroke, bowel disorder, and
 arthritis), and well one derived variable (respiratory condition) to count
 the number of conditions a respondent has.
}
\note{
mood disorder (CCC_280) was not asked to respondents in the 2001 CCHS
 survey cycle. This mean respondents in this cycle will only be able to have
 a maximum of 6 chronic conditions as opposed to 7 for respondents in other
 cycles.
}
\examples{
 # Using multiple_conditions_fun() to generate multiple_conditions in a CCHS
 # cycle.
 
 # multiple_conditions_fun() is specified in variable_details.csv along with
 # the CCHS variables and cycles included.
 
 library(cchsflow)
 conditions_2010 <- rec_with_table(cchs2010, c("CCC_121","CCC_131","CCC_151",
 "CCC_171","CCC_280","resp_condition","CCC_051"))
 
 head(conditions_2010)
 
}
