% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking-validation-constants.R
\docType{data}
\name{PACK_YEARS_CONSTANTS}
\alias{PACK_YEARS_CONSTANTS}
\title{Pack-years calculation constants}
\format{
An object of class \code{list} of length 6.
}
\usage{
PACK_YEARS_CONSTANTS
}
\description{
These constants are used in calculate_pack_years_pumf() and
calculate_pack_years_master() functions. They cannot be expressed in
variable_details.csv because they are calculation parameters, not
recoding bounds.
}
\details{
Evidence base:
- MIN_PACK_YEARS: 100 cigarettes / 7300 = 0.0137 (NHIS/BRFSS "established smoker")
- MIN_PACK_YEARS_ALT: 50 cigarettes / 7300 = 0.007 (youth/experimental smoker)
- MAX_PACK_YEARS: Empirical ceiling from ATBC (162) and Pain & Health (165) studies
}
\seealso{
harmonization-development/smoking/05-pack-years/L2_semantic_mapping.md
}
\keyword{datasets}
