% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{PackYears_fun}
\alias{PackYears_fun}
\title{Smoking pack-years}
\usage{
PackYears_fun(
  SMKDSTY,
  DHHGAGE_cont,
  SMK_09A_B,
  SMKG09C,
  SMKG203_cont,
  SMKG207_cont,
  SMK_204,
  SMK_05B,
  SMK_208,
  SMK_05C,
  SMKG01C_cont,
  SMK_01A
)
}
\arguments{
\item{SMKDSTY}{derived variable that classifies an individual's smoking status.}

\item{DHHGAGE_cont}{continuous age variable.}

\item{SMK_09A_B}{number of years since quitting smoking. Variable asked to former daily smokers who quit <3 years ago.}

\item{SMKG09C}{number of years since quitting smoking. Variable asked to former daily smokers who quit >=3 years ago.}

\item{SMKG203_cont}{age started smoking daily. Variable asked to daily smokers.}

\item{SMKG207_cont}{age started smoking daily. Variable asked to former daily smokers.}

\item{SMK_204}{number of cigarettes smoked per day. Variable asked to daily smokers.}

\item{SMK_05B}{number of cigarettes smoked per day. Variable asked to occasional smokers}

\item{SMK_208}{number of cigarettes smoked per day. Variable asked to former daily smokers}

\item{SMK_05C}{number of days smoked at least one cigarette}

\item{SMKG01C_cont}{age smoked first cigarette}

\item{SMK_01A}{smoked 100 cigarettes in lifetime (y/n)}
}
\value{
value for smoking pack-years in the PackYears_derived variable
}
\description{
Measures an individual smoking pack-years based on various CCHS smoking variables. This is a popular variable
 used by researchers to quantify lifetime exposure to cigarette use.
}
