% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent-time-canada.R
\name{pct_time_fun_A}
\alias{pct_time_fun_A}
\title{Percent time in Canada (Master files)}
\usage{
pct_time_fun_A(DHH_AGE, SDCGCB, SDCDRES)
}
\arguments{
\item{DHH_AGE}{continuous age variable.}

\item{SDCGCB}{whether or not someone was born in Canada (1 - born in Canada,
2 - born outside Canada)}

\item{SDCDRES}{continuous variable for how long someone has lived in Canada
(in years).}
}
\value{
Numeric value between 0 and 100 that represents
 percentage of a respondent's time in Canada
}
\description{
This function creates a derived variable (pct_time_der_A) that
 provides an estimated percentage of the time a person's life was spent in
 Canada. This variant is for Master file data where SDCDRES is continuous
 (actual years) rather than categorical.
}
\examples{
# Using pct_time_fun_A() to generate a value for percent time spent in Canada
# with user inputted values. Let's say you are 27 years old who was born
# outside of Canada and have been living in Canada for 10 years.
# Your estimated percent time spent in Canada can be calculated as follows:

library(cchsflow)
pct_time <- pct_time_fun_A(DHH_AGE = 27, SDCGCB = 2, SDCDRES = 10)

print(pct_time)

}
