% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-validation.R
\name{print.schema_validation_result}
\alias{print.schema_validation_result}
\title{Print Schema Validation Results}
\usage{
\method{print}{schema_validation_result}(x, ...)
}
\arguments{
\item{x}{schema_validation_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Schema Validation Results
}
\keyword{internal}
