% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{recode_cchs_categories}
\alias{recode_cchs_categories}
\title{CCHS variable recoding helper}
\usage{
recode_cchs_categories(variable, recode_map)
}
\arguments{
\item{variable}{Variable to recode}

\item{recode_map}{Named vector mapping old values to new values}
}
\value{
Recoded variable with proper NA handling
}
\description{
CCHS variable recoding helper
}
