% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{resp_condition_fun}
\alias{resp_condition_fun}
\title{resp_condition_fun}
\usage{
resp_condition_fun(age, CCC_091, CCC_031)
}
\arguments{
\item{age}{continuous age variable (DHHGAGE_cont for _p cycles, DHH_AGE for
_m cycles).}

\item{CCC_091}{variable indicating if respondent has COPD, emphysema, or
bronchitis}

\item{CCC_031}{variable indicating if respondent has asthma}
}
\value{
a categorical variable (resp_condition_der) with 3 levels:

 \enumerate{
 \item respondent is over the age of 35 and has a respiratory condition
 \item respondent is under the age of 35 and has a respiratory condition
 \item respondent does not have a respiratory condition
 }
}
\description{
This function is used to create a derived variable
 (resp_condition_der) that determines if a respondent has a respiratory
 condition. It uses the combined COPD, Emphysema, and Bronchitis variable
 (CCC_091) and Asthma (CCC_031) across all CCHS cycles.
}
\examples{
# Using resp_condition_fun() to create values across CCHS cycles
# (2009-2014) resp_condition_fun() is specified in
# variable_details.csv along with the CCHS variables and cycles included.

# To transform resp_condition_der, use rec_with_table() for each CCHS cycle
# and specify resp_condition_der, along with the various respiratory
# variables. Then by using merge_rec_data() you can combine
# resp_condition_der across cycles.

library(cchsflow)

resp2009_2010 <- suppressWarnings(rec_with_table(
  cchs2009_2010_p, c(
    "DHHGAGE_cont", "CCC_091", "CCC_031",
    "resp_condition_der"
  )
))

head(resp2009_2010)

resp2011_2012 <- suppressWarnings(rec_with_table(
  cchs2011_2012_p, c(
    "DHHGAGE_cont", "CCC_091", "CCC_031",
    "resp_condition_der"
  )
))

tail(resp2011_2012)

combined_resp <-
  suppressWarnings(merge_rec_data(resp2009_2010, resp2011_2012))

head(combined_resp)
tail(combined_resp)
}
