% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{RespCondition_fun1}
\alias{RespCondition_fun1}
\title{RespCondition_fun1}
\usage{
RespCondition_fun1(DHHGAGE_cont, CCC_091)
}
\arguments{
\item{DHHGAGE_cont}{continuous age variable.}

\item{CCC_091}{variable indicating if respondent has either COPD or Emphysema}
}
\value{
a categorical variable (RespCondition_derived) with 3 levels:
 1 - respondent is over the age of 35 and has a respiratory condition
 2 - respondent is under the age of 35 and has a respiratory condition
 3 - respondent does not have a respiratory condition
}
\description{
This is one of 3 functions used to create a derived variable (RespCondition_derived) that determines if a respondents has a respirtory condition.
 3 different functions have been created to account for the fact that different respiratory variables are used across CCHS cycles.
 This function is for CCHS cycles (2009-2014) that only use COPD and Emphysema as a combined variable.
}
\seealso{
RespCondition_fun2, RespCondition_fun3
}
