% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-variables.R
\name{RespCondition_fun2}
\alias{RespCondition_fun2}
\title{Respiratory condition variable}
\usage{
RespCondition_fun2(Age_cont, Emphys, COPD, Bronch)
}
\arguments{
\item{Age_cont}{continuous age variable. 
cchsflow variable name: DHHGAGE_cont}

\item{Emphys}{variable indicating if subject has emphysema
cchsflow variable name: CCC_91E}

\item{COPD}{variable indicating if subject has COPD
cchsflow variable name: CCC_91F}

\item{Bronc}{variable indicating if subject has chronic bronchitis
cchsflow variable name: CCC_91A}
}
\value{
a categorical variable (RespCondition_derived) with 3 levels:
1 - subject is over the age of 35 and has a respiratory condition
2 - subject is under the age of 35 and has a respiratory condition
3 - subject does not have a respiratory condition
}
\description{
This function for CCHS cycles that use COPD & Emphysema as separate variables, as well as Bronchitis
}
