% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{RespCondition_fun3}
\alias{RespCondition_fun3}
\title{RespCondition_fun3}
\usage{
RespCondition_fun3(DHHGAGE_cont, CCC_091, CCC_91A)
}
\arguments{
\item{DHHGAGE_cont}{continuous age variable.}

\item{CCC_091}{variable indicating if respondent has either COPD or Emphysema}

\item{CCC_91A}{variable indicating if respondent has chronic bronchitis
cchsflow variable name: CCC_91A}
}
\value{
a categorical variable (RespCondition_derived) with 3 levels:
 1 - respondent is over the age of 35 and has a respiratory condition
 2 - respondent is under the age of 35 and has a respiratory condition
 3 - respondent does not have a respiratory condition
}
\description{
This is one of 3 functions used to create a derived variable (RespCondition_derived) that determines if a respondents has a respirtory condition.
 This function for CCHS cycles (2001-2003) that use COPD and Emphysema as a combined variable, as well as Bronchitis
}
\seealso{
RespCondition_fun1, RespCondition_fun2
}
