% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{sedentary_activity2_fun}
\alias{sedentary_activity2_fun}
\title{Total hours - sedentary activity excluding reading (age 20+)}
\usage{
sedentary_activity2_fun(age, SAC_1, SAC_2, SAC_3)
}
\arguments{
\item{age}{Continuous age (DHHGAGE_cont for PUMF, DHH_AGE for Master)}

\item{SAC_1}{Hours on computer}

\item{SAC_2}{Hours playing video games}

\item{SAC_3}{Hours watching TV or videos}
}
\value{
Total sedentary hours (excl. reading) for age 20+, or tagged NA
}
\description{
Calculates total hours spent doing sedentary activities
 excluding reading: computer use, video games, and watching TV/videos.
 Time spent at school or work is excluded. Limited to ages 20+.
}
\examples{
sedentary_activity2_fun(25, 5, 3, 10)  # 18 hours
sedentary_activity2_fun(18, 5, 3, 10)  # NA(a) - under 20

}
