% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{sedentary_activity_fun}
\alias{sedentary_activity_fun}
\title{Total hours - sedentary activity (age 20+)}
\usage{
sedentary_activity_fun(age, SAC_1, SAC_2, SAC_3, SAC_4)
}
\arguments{
\item{age}{Continuous age (DHHGAGE_cont for PUMF, DHH_AGE for Master)}

\item{SAC_1}{Hours on computer}

\item{SAC_2}{Hours playing video games}

\item{SAC_3}{Hours watching TV or videos}

\item{SAC_4}{Hours reading}
}
\value{
Total sedentary hours for age 20+, or tagged NA
}
\description{
Calculates total hours spent doing sedentary activities:
 computer use, video games, watching TV/videos, and reading.
 Time spent at school or work is excluded. Limited to ages 20+.
}
\examples{
sedentary_activity_fun(25, 5, 3, 10, 7)  # 25 hours
sedentary_activity_fun(18, 5, 3, 10, 7)  # NA(a) - under 20

}
