% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-utils.R
\name{set_data_labels}
\alias{set_data_labels}
\title{Set data labels}
\usage{
set_data_labels(data_to_label, variable_details, variables_sheet = NULL)
}
\arguments{
\item{variable_details}{A dataframe containing the details of each variable,
with category labels in the 'catLabel' column.}

\item{variables_sheet}{(Optional) A dataframe containing variable labels in
the 'label' column.}

\item{data_to_lab.el}{A dataframe of CCHS data that lacks labels.}
}
\value{
The dataframe with labelled CCHS variables.
}
\description{
Set variable labels for a working CCHS dataframe.  The function
extracts variables from the provided dataframe and assigns labels according
the provided variables and variable_details dataframes.
in the passed dataframes
}
\examples{
library(cchsflow)
library(sjlabelled) # used to get labels in the example.

bmi2001 <- rec_with_table(
  cchs2001_p, c(
    "HWTGHTM",
    "HWTGWTK", "HWTGBMI_der"
  )
)

bmi2003 <- rec_with_table(
  cchs2003_p, c(
    "HWTGHTM",
    "HWTGWTK", "HWTGBMI_der"
  )
)

combined_bmi <- bind_rows(bmi2001, bmi2003)

get_label(combined_bmi)

labeled_combined_data <- set_data_labels(
  combined_bmi,
  variable_details,
  variables
)

get_label(labeled_combined_data)
}
