% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{SMK09AB_fun}
\alias{SMK09AB_fun}
\title{Years since stopped smoking for former daily smokers (categorical - version 2)}
\usage{
SMK09AB_fun(SPU_25I)
}
\arguments{
\item{SPU_25I}{Time since stopped smoking for former daily smokers}
}
\value{
the CCHS derived variable SMK_09A_B is a categorical variable
}
\description{
This function creates a derived variable (SMK_09A_B)
 that calculates the approximate time since a former daily smoker has quit smoking
 with categories: less than a year, 1 to <2 years, 2 to <3 years, 3 years or more
}
