% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{SMK09AI_fun}
\alias{SMK09AI_fun}
\title{Years since stopped smoking for former daily smokers (categorical - version 1)}
\usage{
SMK09AI_fun(SPU_25I)
}
\arguments{
\item{SPU_25I}{Time since stopped smoking for former daily smokers}
}
\value{
the CCHS derived variable SMK_09A_I is a categorical variable
}
\description{
This function creates a derived variable (SMK_09A_I)
 that calculates the approximate time since a former daily smoker has quit smoking
 with categories: half a year, one and a half years, two and a half years, 4 years or more
}
