% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{SMK09C_fun}
\alias{SMK09C_fun}
\title{Years since stopped smoking for former daily smokers (categorical - version 3)}
\usage{
SMK09C_fun(SPU_25I)
}
\arguments{
\item{SPU_25I}{Time since stopped smoking for former daily smokers}
}
\value{
the CCHS derived variable SMKG09C_I is a categorical variable
}
\description{
This function creates a derived variable (SMKG09C)
 that calculates the approximate years since a former daily smoker has quit smoking
 with categories: 3 to 5 years, 6 to 10 years, 11 years or more
}
