% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{SMKG203I_fun}
\alias{SMKG203I_fun}
\title{Age started to smoke daily - daily smoker (modern CCHS 2015+)}
\usage{
SMKG203I_fun(SMK_005, SMKG040_I)
}
\arguments{
\item{SMK_005}{Type of smoker presently}

\item{SMKG040_I}{Age started to smoke daily - daily/former daily smoker (continuous)}
}
\value{
the CCHS derived variable SMKG203_cont is a continuous variable
}
\description{
This function creates a derived variable (SMKG203_cont)
 that provides the age started to smoke daily for current daily smokers.
}
