% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{SMKG207I_fun}
\alias{SMKG207I_fun}
\title{Age started to smoke daily - former daily smoker (modern CCHS 2015+)}
\usage{
SMKG207I_fun(SMK_030, SMKG040_I)
}
\arguments{
\item{SMK_030}{Smoked daily - lifetime (occasional/former smoker)}

\item{SMKG040_I}{Age started to smoke daily - daily/former daily smoker (continuous)}
}
\value{
the CCHS derived variable SMKG207_cont is a continuous variable
}
\description{
This function creates a derived variable (SMKG207_cont)
 that provides the age started to smoke daily for former daily smokers.
}
