% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{SPU25_fun}
\alias{SPU25_fun}
\title{Time since stopped smoking for former daily smokers (cycle 2022)}
\usage{
SPU25_fun(SPU_25A_I, SPU_25B_I, ADM_MOI_I, ADM_YOI_I)
}
\arguments{
\item{SPU_25A_I}{stopped smoking daily - month (former daily smoker)}

\item{SPU_25B_I}{stopped smoking daily - year (former daily smoker)}

\item{ADM_MOI_I}{month of interview}

\item{ADM_YOI_I}{year of interview}
}
\value{
the CCHS derived variable SPU_25I is a continuous variable
}
\description{
This function creates a derived variable (SPU_25I)
 that calculates the approximate time since a former daily smoker has quit smoking based
 on 2022 CCHS variables.
}
