% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_details}
\alias{variable_details}
\title{variable_details.csv}
\value{
\item{variable_details}{a data frame}
}
\description{
This dataset provides details on how variables are specified in cchsflow.
}
\details{
See the below link for more details about how the worksheet is structured
\url{https://big-life-lab.github.io/cchsflow/articles/variablesDetails.html}
}
\examples{
data(variable_details)
str(variable_details)
}
\keyword{datasets}
