% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables}
\alias{variables}
\title{variables.csv}
\value{
\item{variables}{a data frame}
}
\description{
This dataset lists all the variables that are present in cchsflow.
}
\details{
See \href{https://big-life-lab.github.io/cchsflow/articles/variablesSheet.html}{here} for more
details about how the worksheet is structured
}
\examples{
data(variables)
str(variables)
}
\keyword{datasets}
