% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{video_game_age1_fun}
\alias{video_game_age1_fun}
\title{Time spent playing video games (age 20+)}
\usage{
video_game_age1_fun(DHHGAGE_cont, SAC_2D)
}
\arguments{
\item{DHHGAGE_cont}{Continuous age}

\item{SAC_2D}{No./hours in past week playing video games for CCHS 2015-2016}
}
\value{
value for hours/week playing video games
}
\description{
This function creates a derived variable (SAC_2_A)
 that calculates the total hours in a week spent playing video games for age
 20+.
}
\examples{
# Using video_game_age1_fun() to create values for hours/week playing video 
# games across CCHS cycles.
# video_game_age1_fun() is specified in variable_details.csv along with 
# the CCHS variables and cycles included.

# To transform SAC_2_A across cycles, use rec_with_table() for each
# CCHS cycle and specify SAC_2_A, along with the required variables
# variables. For CCHS 2009-2014, only DHHGAGE_cont and SAC_2_A need to be 
# specified in the transformation. For CCHS 2015-2016, DHHGAGE_cont, 
# SAC_2_A, and the associated variables for SAC_2D need to be 
# specified in the transformation. Then by using merge_rec_data(), you can 
# combine SAC_2_A across cycles.

library(cchsflow)

vg_A_2013_2014 <- rec_with_table(
  cchs2013_2014_p, c(
    "DHHGAGE_cont", "SAC_2_A"
  )
)

head(vg_A_2013_2014)

vg_A_2015_2016 <- rec_with_table(
  cchs2015_2016_p, c(
    "SAC_2D_A","SAC_2D_B", "SAC_2D", "DHHGAGE_cont", "SAC_2_A"
  )
)

tail(vg_A_2015_2016)

combined_vg_A <- suppressWarnings(merge_rec_data(vg_A_2013_2014,
vg_A_2015_2016))

head(combined_vg_A)
tail(combined_vg_A)
}
