% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{video_game_fun}
\alias{video_game_fun}
\title{Video games (physical and non-physical activity)}
\usage{
video_game_fun(SAC_2D_A, SAC_2D_B)
}
\arguments{
\item{SAC_2D_A}{No./hours in past week playing video games - physical 
activity for CCHS 2015-2016}

\item{SAC_2D_B}{No./hours in past week playing video games - non-physical 
activity for CCHS 2015-2016}
}
\value{
value for hours/week playing video games
}
\description{
This function creates a derived variable (SAC_2D)
 that calculates the total hours in a week spent playing video games (both
 physical and non-physical activity).
}
\examples{
# Using video_game_fun() to create values for hours/week playing video games 
# across CCHS cycles.
# video_game_fun() is specified in variable_details.csv along with 
# the CCHS variables and cycles included.

# To transform SAC_2D, use rec_with_table() for each
# CCHS cycle and specify SAC_2D, along with SAC_2D_A and SAC_2D_B
# variables.Then by using merge_rec_data(), you can combine 
# SAC_2D for CCHS 2015-2016.

library(cchsflow)

vg2015_2016 <- rec_with_table(
  cchs2015_2016_p, c(
    "SAC_2D_A","SAC_2D_B", "SAC_2D"
  )
)

head(vg2015_2016)
tail(vg2015_2016)
}
