% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{weekly_screen_time_age2_fun}
\alias{weekly_screen_time_age2_fun}
\title{Weekly leisure screen time (age <20)}
\usage{
weekly_screen_time_age2_fun(DHHGAGE_cont, SBE_010)
}
\arguments{
\item{DHHGAGE_cont}{Continuous age}

\item{SBE_010}{Daily screen time (hours/day) outside of school and work day for
CCHS 2017-2018}
}
\value{
value for screen time (hours/week) outside of school and work day
}
\description{
This function creates a derived variable (weekly_screen_B)
 that calculates the total hours in a week spent using a screen outside of 
 school and/or work. This is limited to age less than 20.
}
