% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{weekly_screen_time_fun}
\alias{weekly_screen_time_fun}
\title{Weekly leisure screen time}
\usage{
weekly_screen_time_fun(SBE_010)
}
\arguments{
\item{SBE_010}{Daily screen time (hours/day) outside of school and work}
}
\value{
Weekly screen time (hours/week) or tagged NA
}
\description{
Calculates total weekly hours of screen time outside of
 school and/or work from daily screen time.
}
\examples{
weekly_screen_time_fun(2)  # 14 hours/week
weekly_screen_time_fun(0)  # 0

}
