% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{weekly_screen_time_youth_fun}
\alias{weekly_screen_time_youth_fun}
\title{Weekly leisure screen time - youth (age under 20)}
\usage{
weekly_screen_time_youth_fun(age, SBE_010)
}
\arguments{
\item{age}{Continuous age (DHHGAGE_cont for PUMF, DHH_AGE for Master)}

\item{SBE_010}{Daily screen time (hours/day) outside of school and work}
}
\value{
Weekly screen time (hours/week) for age under 20, or tagged NA
}
\description{
Calculates total weekly screen time outside of school/work
 for respondents aged under 20. Returns NA(a) for age 20 and over.
}
