% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{adjusted_bmi_fun}
\alias{adjusted_bmi_fun}
\title{Adjusted Body Mass Index (BMI) derived variable}
\usage{
adjusted_bmi_fun(
  DHH_SEX,
  HWTGHTM,
  HWTGWTK,
  min_HWTGHTM = 0.82,
  max_HWTGHTM = 2.5,
  min_HWTGWTK = 25,
  max_HWTGWTK = 145,
  BMI_min = 8.07,
  BMI_max = 137.46
)
}
\arguments{
\item{DHH_SEX}{CCHS variable for sex (1=male, 2=female)}

\item{HWTGHTM}{CCHS variable for height (in meters)}

\item{HWTGWTK}{CCHS variable for weight (in kilograms)}

\item{min_HWTGHTM, max_HWTGHTM, min_HWTGWTK, max_HWTGWTK, BMI_min, BMI_max}{Validation parameters}
}
\value{
numeric value for adjusted BMI. Returns tagged_na("b") for invalid measurements.
}
\description{
BMI with sex-specific corrections for self-reporting bias.
}
