% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adl.R, R/adl_score_6.R
\name{adl_fun_B}
\alias{adl_fun_B}
\title{This newly derived variable (ADL_der_B) uses ADL_01 to ADL_06 which are
 consistent across survey cycles from 2003-2014. ADL_der_B is recommended when using multiple
 survey cycles.}
\usage{
adl_fun_B(ADL_01, ADL_02, ADL_03, ADL_04, ADL_05, ADL_06)

adl_fun_B(ADL_01, ADL_02, ADL_03, ADL_04, ADL_05, ADL_06)
}
\arguments{
\item{ADL_01}{Needs help preparing meals}

\item{ADL_02}{Needs help getting to appointments/errands}

\item{ADL_03}{Needs help doing housework}

\item{ADL_04}{Needs help doing personal care}

\item{ADL_05}{Needs help moving inside house}

\item{ADL_06}{Needs help doing finances}
}
\value{
A derived variable (ADL_der_B) with 2 categories:
 \enumerate{
  \item - Needs help with tasks
  \item - Does not need help with tasks
 }
}
\description{
This derived variable (ADL_der) is based on the CCHS derived
 variable ADLF6R which flags respondents who need help with tasks based on
 their response to the various activities of daily living (ADL) variables.
}
\details{
The CCHS derived variable ADLF6R uses different ADL variables across
 the various CCHS survey cycles. This newly derived variable (ADL_der) uses
 ADL variables that are consistent across CCHS cycles.
 
 In the 2001 CCHS survey cycle, the ADLF6R variable examines the following
 ADL variables:
 \enumerate{
  \item ADL_01 - Needs help preparing meals
  \item ADL_02 - Needs help getting to appointments/errands
  \item ADL_03 - Needs help doing housework
  \item ADL_04 - Needs help doing personal care
  \item ADL_05 - Needs help moving inside house
  \item ADL_07 - Needs help doing heavy household chores
 }  

 In the 2003-2005 CCHS survey cycles, the ADLF6R variable examines the
 following ADL variables:
 \enumerate{
  \item ADL_01 - Needs help preparing meals
  \item ADL_02 - Needs help getting to appointments/errands
  \item ADL_03 - Needs help doing housework
  \item ADL_04 - Needs help doing personal care
  \item ADL_05 - Needs help moving inside house
  \item ADL_06 - Needs help doing finances
  \item ADL_07 - Needs help doing heavy household chores
 }
 
 In the 2007-2014 CCHS survey cycles, the ADLF6R variable examines the
 following ADL variables:
 \enumerate{
  \item ADL_01 - Needs help preparing meals
  \item ADL_02 - Needs help getting to appointments/errands
  \item ADL_03 - Needs help doing housework
  \item ADL_04 - Needs help doing personal care
  \item ADL_05 - Needs help moving inside house
  \item ADL_06 - Needs help doing finances
 }
 
 This newly derived variable (ADL_der_B) uses ADL_01 to ADL_06 which are
 consistent across survey cycles from 2003-2014. ADL_der_B is recommended when using multiple
 survey cycles.
}
\examples{
# Using adl_fun_B() to create ADL_der_B values across CCHS cycles
# adl_fun_B() is specified in variable_details.csv along with the
# CCHS variables and cycles included.

# To transform ADL_der_B, use rec_with_table() for each CCHS cycle
# and specify ADL_der_B, along with the various ADL variables.
# Then by using merge_rec_data() you can combine ADL_der_B across cycles.

library(cchsflow)
adl2001 <- rec_with_table(
  cchs2001_p, c(
    "ADL_01", "ADL_02", "ADL_03", "ADL_04", "ADL_05", "ADL_der_B"
  )
)

head(adl2001)

adl2009_2010 <- rec_with_table(
  cchs2009_2010_p, c(
    "ADL_01", "ADL_02", "ADL_03", "ADL_04", "ADL_05", "ADL_06", "ADL_der_B"
  )
)

tail(adl2009_2010)

combined_adl <- merge_rec_data(adl2001, adl2009_2010)

head(combined_adl)

tail(combined_adl)

# Using adl_fun_B() to generate to ADL_der_B based on user inputted values.
# 
# Let's say you do not need help preparing meals, you need help getting to
# appointments or errands, you need help doing housework, do not need help
# doing personal care, do not need help moving inside the house, and do not 
# need help doing fiances. Using adl_fun_B() we can check if you need help 
# doing tasks

ADL_der_B <- adl_fun_B(2, 1, 1, 2, 2, 2)

print(ADL_der_B)

}
