% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-categorical.R
\name{age_cat_fun}
\alias{age_cat_fun}
\title{Derived categorical age}
\usage{
age_cat_fun(DHHGAGE_cont)
}
\arguments{
\item{DHHGAGE_cont}{continuous age variable}
}
\description{
This is a derived categorical age variable that groups various
 age categories across all CCHS cycles. This is based on the continuous age
 variable (DHHGAGE_cont) that is harmonious across all CCHS cycles.

 The categories of this variable are based on the age groupings seen in CCHS
 cycles from 2007 to 2014. The age breakdown is as follows:

 \enumerate{
   \item 12 to 14 years
   \item 15 to 17 years
   \item 18 to 19 years
   \item 20 to 24 years
   \item 25 to 29 years
   \item 30 to 34 years
   \item 35 to 39 years
   \item 40 to 44 years
   \item 45 to 49 years
   \item 50 to 54 years
   \item 55 to 59 years
   \item 60 to 64 years
   \item 65 to 69 years
   \item 70 to 74 years
   \item 75 to 79 years
   \item 80 years or more
 }
}
\details{
The categories in the grouped age variable (DHHGAGE) vary between
 CCHS cycles. As such, a continous age variable (DHHGAGE_cont) was created
 that harmonized age across all CCHS cycle by taking the midpoint of each
 age category.
}
