% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi_fun}
\alias{bmi_fun}
\title{Body Mass Index (BMI) derived variable}
\usage{
bmi_fun(
  HWTGHTM,
  HWTGWTK,
  min_HWTGHTM = 0.82,
  max_HWTGHTM = 2.5,
  min_HWTGWTK = 25,
  max_HWTGWTK = 145,
  BMI_min = 8.07,
  BMI_max = 137.46
)
}
\arguments{
\item{HWTGHTM}{CCHS variable for height (in meters)}

\item{HWTGWTK}{CCHS variable for weight (in kilograms)}

\item{min_HWTGHTM, max_HWTGHTM, min_HWTGWTK, max_HWTGWTK, BMI_min, BMI_max}{Validation parameters}
}
\value{
numeric value for BMI. Returns tagged_na("b") for invalid measurements.
}
\description{
This function creates a harmonized BMI variable across CCHS cycles.
}
\note{
v2.2.0, last updated: 2025-06-30, status: active, Note: Enhanced validation for new BMI variables (HWTDBMI, HWTDHTM, HWTDWTK)
}
