% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-legacy.R
\name{bmi_fun_cat}
\alias{bmi_fun_cat}
\title{Categorical BMI (international standard)}
\usage{
bmi_fun_cat(...)
}
\arguments{
\item{HWTGBMI_der}{derived BMI variable}
}
\value{
BMI category (1=underweight, 2=normal, 3=overweight, 4=obese)
}
\description{
Categorizes BMI into 4 standard categories.
}
