% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{calculate_pack_years}
\alias{calculate_pack_years}
\title{Pack-years calculation helper}
\usage{
calculate_pack_years(cigarettes_per_day, years_smoked, min_pack_years = 0.0137)
}
\arguments{
\item{cigarettes_per_day}{Number of cigarettes smoked per day}

\item{years_smoked}{Number of years smoking}

\item{min_pack_years}{Minimum pack-years value (default 0.0137)}
}
\value{
Pack-years calculation with minimum threshold
}
\description{
Pack-years calculation helper
}
\keyword{internal}
