% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking-status.R
\name{calculate_SMKDSTY_original}
\alias{calculate_SMKDSTY_original}
\title{Derive 6-category smoking status (deprecated alias)}
\usage{
calculate_SMKDSTY_original(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{calculate_SMKDSTY_cat6}}.}
}
\description{
Deprecated. Use \code{\link{calculate_SMKDSTY_cat6}} instead.
}
