% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent-time-canada.R
\name{categorize_pct_time}
\alias{categorize_pct_time}
\title{Categorize percent time in Canada}
\usage{
categorize_pct_time(pct_time_der, output_format = "tagged_na")
}
\arguments{
\item{pct_time_der}{Derived continuous percent time in Canada.
See \code{\link{calculate_pct_time}} for documentation on how variable
was derived.}

\item{output_format}{Output missing data format: "tagged_na" (default) or "original".}
}
\value{
Integer 1-10 for categorical percent time in Canada,
  or tagged NA for missing/not applicable.
}
\description{
Categorizes the derived percent time in Canada variable
 (pct_time_der) into 10 percent intervals for pct_time_der_cat10.
}
\examples{
# Scalar usage
categorize_pct_time(55.5)

# Vector usage
categorize_pct_time(c(5, 25, 55, 85, 100))

}
