% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2005_p}
\alias{cchs2005_p}
\title{2005 CCHS PUMF subset data (200 respondents)}
\source{
\url{http://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=22642}
}
\value{
\item{cchs2005_p}{a data frame}
}
\description{
This is a subset of 200 observations from the 2005 cycle of the Canadian
Community Health Survey (CCHS) Public Use Microdata file (PUMF) dataset.
The CCHS survey is conducted by Statistics Canada.
}
\details{
See \href{https://www.statcan.gc.ca/eng/reference/licence}{here} for the open
license. Source from Statistics Canada, Canadian Community Health Survey 
PUMF, accessed Jan 2020. Reproduced and distributed on an "as is" basis with 
the permission of Statistics Canada.

Long name: cchs-82M0013-E-2005-c3-1-main-file

DDI: \url{https://osf.io/35mhq/}

Additional documentation (PDFs): \url{https://osf.io/hkuy3/}
}
\examples{
data(cchs2005_p)
str(cchs2005_p)
}
\keyword{datasets}
