% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2007_2008_p}
\alias{cchs2007_2008_p}
\title{2007-2008 CCHS data}
\source{
\url{http://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=29539}
}
\value{
\item{cchs2007_2008_p}{a data frame}
}
\description{
This is a subset of 200 observations from the 2007-2008 cycle of the Canadian
Community Health Survey (CCHS) Public Use Microdata file (PUMF) dataset.
The CCHS survey is conducted by Statistics Canada.
}
\details{
See \href{https://www.statcan.gc.ca/eng/reference/licence}{here} for the open
license.

Long name: cchs-E-2007-2008-AnnualComponent

DDI: \url{https://osf.io/emzsp/}

Additional documentation (PDFs): \url{https://osf.io/hkuy3/}
}
\examples{
data(cchs2007_2008_p)
str(cchs2007_2008_p)
}
\keyword{datasets}
