% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2010}
\alias{cchs2010}
\title{2010 CCHS data}
\source{
\url{http://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=81424}
}
\value{
\item{cchs2010}{a data frame}
}
\description{
This is a subset of 200 observations from the 2010 cycle of the Canadian Community Health Survey 
(CCHS) Public Use Microdata file (PUMF) dataset. 
The CCHS survey is conducted by Statistics Canada.
}
\details{
See the open license \href{https://www.statcan.gc.ca/eng/reference/licence}{here}.

Long name: CCHS-82M0013-E-2010-AnnualComponent

DDI: \url{https://github.com/Big-Life-Lab/cchsflow/blob/master/inst/extdata/CCHS_DDI/CCHS-82M0013-E-2010-AnnualComponent.xml}

Additional documentation (PDFs): \url{https://osf.io/hkuy3/}
}
\examples{
data(cchs2010)
str(cchs2010)
}
\keyword{datasets}
