% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cchs2012_s}
\alias{cchs2012_s}
\title{2012 CCHS synthetic subset data (200 respondents)}
\source{
\url{https://www.statcan.gc.ca/en/statistical-programs/document/3226_D56_T9_V1}
}
\value{
\item{cchs2012_s}{a data frame}
}
\description{
This is a subset of 200 observations from the 2012 cycle of the Canadian
Community Health Survey (CCHS) synthetic dataset.
The CCHS survey is conducted by Statistics Canada.
}
\details{
\strong{NOTE:} this subset of respondents may also be in the 2012 synthetic
subset. Please see the "CCHS datasets that overlap each other" article to see
how the two datasets contain overlap.

See \href{https://www.statcan.gc.ca/en/reference/licence}{here} for the open
license. Source from Statistics Canada, Canadian Community Health Survey 
PUMF, accessed May 2022. Reproduced and distributed on an "as is" basis with 
the permission of Statistics Canada.

Long name: synthetic-CCHS-E-2012-AnnualComponent_F1

DDI: \url{https://osf.io/q8g7y/}

Additional documentation (PDFs): \url{https://osf.io/hkuy3/}
}
\examples{
data(cchs2012_s)
str(cchs2012_s)
}
\keyword{datasets}
