% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.check_column_order}
\alias{.check_column_order}
\title{Check the columns order in a worksheet}
\usage{
.check_column_order(csv_data, expected_columns, error_ctx)
}
\arguments{
\item{csv_data}{A data.frame containing the worksheet rows}

\item{expected_columns}{The worksheet column in their expected order}

\item{error_ctx}{Information used when creating the error object. A named
list with the following fields:
* file_type: The type of worksheet the CSV contains. Can be "variables" or
  "variable_details".
* file_path: The file path to the worksheet}
}
\value{
The list of column order errors found in the worksheet
}
\description{
Check the columns order in a worksheet
}
