% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.check_excessive_quoting}
\alias{.check_excessive_quoting}
\title{Check a worksheet for excessive quoting}
\usage{
.check_excessive_quoting(parsed_csv, error_ctx)
}
\arguments{
\item{parsed_csv}{Parsed worksheet contents (list of rows)}

\item{error_ctx}{Information used when creating the error object. A named
list with the following fields:
* file_type: The type of worksheet the CSV contains. Can be "variables" or
  "variable_details".
* file_path: The file path to the worksheet}
}
\value{
list of errors
}
\description{
Check a worksheet for excessive quoting
}
