% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.check_line_endings}
\alias{.check_line_endings}
\title{Check whether a worksheet has the correct line endings}
\usage{
.check_line_endings(parsed_csv, error_ctx)
}
\arguments{
\item{error_ctx}{Information used when creating the error object. A named
list with the following fields:
* file_type: The type of worksheet the CSV contains. Can be "variables" or
  "variable_details".
* file_path: The file path to the worksheet}

\item{csv_text}{A string containing the worksheet contents}
}
\value{
The list of line ending errors found in the worksheet
}
\description{
Check whether a worksheet has the correct line endings
}
