% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.create_column_order_error}
\alias{.create_column_order_error}
\title{Create an error for when the worksheet columns are in the wrong order}
\usage{
.create_column_order_error(
  file_type,
  file_path,
  expected_column,
  col_num,
  actual_column
)
}
\arguments{
\item{file_type}{The type of worksheet. Can be "variables" or
"variable_details".}

\item{file_path}{Path to the worksheet}

\item{expected_column}{The column expected at the offending position}

\item{col_num}{The position of the column with the wrong header}

\item{actual_column}{The actual column value}
}
\value{
A named list
}
\description{
Create an error for when the worksheet columns are in the wrong order
}
