% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.create_excessive_quoting_error}
\alias{.create_excessive_quoting_error}
\title{Create an error for when the worksheet has excessive quoting}
\usage{
.create_excessive_quoting_error(
  file_type,
  file_path,
  row_num,
  col_num,
  cell_value
)
}
\arguments{
\item{file_path}{Path to the worksheet}

\item{row_num}{Row number with excessive quotes}

\item{col_num}{Column number with excessive quotes}

\item{cell_value}{Value of the cell with excessive quotes}

\item{file_type:}{The type of worksheet. Can be "variables" or
"variable_details".}
}
\value{
A named list
}
\description{
Create an error for when the worksheet has excessive quoting
}
