% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.create_file_not_found_error}
\alias{.create_file_not_found_error}
\title{Create the error object for when the worksheet could not be found}
\usage{
.create_file_not_found_error(file_type, file_path)
}
\arguments{
\item{file_path}{The invalid path}

\item{file_type:}{The type of worksheet. Can be "variables" or
"variable_details".}
}
\value{
A named list
}
\description{
Create the error object for when the worksheet could not be found
}
