% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.create_invalid_csv_error}
\alias{.create_invalid_csv_error}
\title{Create the error for when the worksheet is not valid CSV}
\usage{
.create_invalid_csv_error(file_type, file_path, error_message)
}
\arguments{
\item{file_path}{Path to the worksheet}

\item{error_message}{Reason(s) for why the worksheet is invalid CSV}

\item{file_type:}{The type of worksheet. Can be "variables" or
"variable_details".}
}
\value{
A named list
}
\description{
Create the error for when the worksheet is not valid CSV
}
