% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.create_line_ending_crlf_error}
\alias{.create_line_ending_crlf_error}
\title{Create an error for when the worksheet has invalid line endings}
\usage{
.create_line_ending_crlf_error(file_type, file_path, row_num)
}
\arguments{
\item{file_type}{The type of worksheet. Can be "variables" or
"variable_details".}

\item{file_path}{Path to the worksheet}

\item{row_num}{Index of the row with the invalid line ending}
}
\value{
A named list
}
\description{
Create an error for when the worksheet has invalid line endings
}
