% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.create_trailing_empty_columns_error}
\alias{.create_trailing_empty_columns_error}
\title{Create an error for when the worksheet has trailing empty columns}
\usage{
.create_trailing_empty_columns_error(file_type, file_path, col_num)
}
\arguments{
\item{file_path}{Path to the worksheet}

\item{col_num}{Position of the trailing empty column}

\item{file_type:}{The type of worksheet. Can be "variables" or
"variable_details".}
}
\value{
A named list
}
\description{
Create an error for when the worksheet has trailing empty columns
}
