% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-worksheet.R
\name{.create_unsorted_rows_error}
\alias{.create_unsorted_rows_error}
\title{Create an error for when the worksheet rows are unsorted}
\usage{
.create_unsorted_rows_error(file_type, file_path, id_column_name)
}
\arguments{
\item{file_path}{Path to the worksheet}

\item{id_column_name}{Name of the column that should be sorted}

\item{file_type:}{The type of worksheet. Can be "variables" or
"variable_details".}
}
\value{
A named list
}
\description{
Create an error for when the worksheet rows are unsorted
}
