% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression.R
\name{DPSDSF}
\alias{DPSDSF}
\title{Derived Depression Scale - Short Form Score}
\usage{
DPSDSF(DPSDSF)
}
\arguments{
\item{DPSDSF}{cchsflow variable name for derived depression scale.}
}
\description{
\strong{NOTE:} this is not a function.

 This is a continuous variable derived by Statistics Canada that assesses the
 level of depression of respondents who have identified that they have felt
 depressed or loss of interest within the last two weeks. This variable is
 scaled from 0 to 8, with 0 indicating a respondent has not felt depressed or
 loss of interest, and 8 representing the highest level of depression.
}
\details{
The derivation of this variable is based on the work of Kessler &
 Mroczek from the University of Michigan. For more details on the items used
 and how the variable was derived click \href{https://osf.io/auqwh/}{here}.
 
 This variable is present in every CCHS cycle used in cchsflow, and
 how it was derived remains consistent.
}
\examples{
library(cchsflow)
 ?DPSDSF

}
\seealso{
\code{\link{DPSDPP}}
}
