% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{ifelse2}
\alias{ifelse2}
\title{ifelse2}
\usage{
ifelse2(x, a, b)
}
\arguments{
\item{x}{A logical argument}

\item{a}{value if `x` is `TRUE`}

\item{b}{value if `x` is `FALSE`}
}
\value{
a or b based on the evaluation of x
}
\description{
Custom ifelse function that evaluates missing (NA) values
}
\details{
If the logical argument (x) compares to a value that is `NA`, it is set to `FALSE`
}
\examples{
age <- 12
status <- ifelse2((age < 18), "child", "invalid age")
print(status)

age <- NA
status <- ifelse2((age < 18), "child", "invalid age")
print(status)
}
