% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-with-table.R
\name{is_equal}
\alias{is_equal}
\title{is equal}
\usage{
is_equal(v1, v2)
}
\arguments{
\item{v1}{variable 1}

\item{v2}{variable 2}
}
\description{
Function to compare even with NA present
This function returns TRUE wherever elements are the same, including NA's,
and false everywhere else.
}
