% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{pack_years_fun_cat}
\alias{pack_years_fun_cat}
\title{Categorical smoking pack-years}
\usage{
pack_years_fun_cat(pack_years_der)
}
\arguments{
\item{pack_years_der}{Smoking pack-years}
}
\description{
This function creates a categorical derived variable
(pack_years_cat) that categorizes smoking pack-years (pack_years_der).
}
\details{
The categories were based on the Cardiovascular Disease Population
Risk Tool (Douglas Manuel et al. 2018).
}
