% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{pack_years_fun_cat}
\alias{pack_years_fun_cat}
\title{Categorical smoking pack-years}
\usage{
pack_years_fun_cat(pack_years_der)
}
\arguments{
\item{pack_years_der}{derived variable that calculates smoking pack-years}
}
\value{
value for pack year categories in the pack_years_cat variable.
}
\description{
This function creates a categorical derived variable
(pack_years_cat) that categorizes smoking pack-years (pack_years_der).
}
\note{
v2.2.0, last updated: 2025-06-30, status: active, Note: Improved categorization with tidyverse patterns
}
