% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking-variables.R
\name{PackYears_fun}
\alias{PackYears_fun}
\title{Smoking pack-years}
\usage{
PackYears_fun(
  TypeOfSmoker,
  Age_cont,
  stpd,
  stpdy,
  agecigd,
  agecigfd,
  cigdayd,
  cigdayo,
  cigdayf,
  dayocc,
  agec1,
  s100
)
}
\arguments{
\item{TypeOfSmoker}{derived variable that classifies an individual's smoking status. 
cchsflow variable name: SMKDSTY}

\item{Age_cont}{continuous age variable. 
cchsflow variable name: DHHGAGE_cont}

\item{stpd}{number of years since quitting smoking. Variable asked to former daily smokers who quit <3 years ago. 
cchsflow variable name: SMK_09A_B}

\item{stpdy}{number of years since quitting smoking. Variable asked to former daily smokers who quit >=3 years ago. 
cchsflow variable name: SMKG09C}

\item{agecigd}{age started smoking daily. Variable asked to daily smokers. cchsflow variable name: SMKG203_cont}

\item{agecigfd}{age started smoking daily. Variable asked to former daily smokers. 
cchsflow variable name: SMKG207_cont}

\item{cigdayd}{number of cigarettes smoked per day. Variable asked to daily smokers.
cchsflow variable name: SMK_204}

\item{cigdayo}{number of cigarettes smoked per day. Variable asked to occasional smokers
cchsflow variable name: SMK_05B}

\item{cigdayf}{number of cigarettes smoked per day. Variable asked to former daily smokers
cchsflow variable name: SMK_208}

\item{dayocc}{number of days smoked at least one cigarette
cchsflow variable name: SMK_05C}

\item{agec1}{age smoked first cigarette
cchsflow variable name: SMKG01C_cont}

\item{s100}{smoked 100 cigarettes in lifetime (y/n)
cchsflow variable name: SMK_01A}
}
\value{
value for smoking pack-years in the PackYears_derived variable
}
\description{
Measures an individual smoking pack-years based on various CCHS smoking variables
}
