% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent-time-canada.R
\name{Pct_time_fun}
\alias{Pct_time_fun}
\title{Percent time in Canada}
\usage{
Pct_time_fun(Age_cont, BirthCountry, TimeCanada)
}
\arguments{
\item{Age_cont}{continuous age variable. 
cchsflow variable name: DHHGAGE_cont}

\item{BirthCountry}{whether or not someone was born in Canada 
cchsflow variable name: SDCGCBG}

\item{TimeCanada}{how long someone has lived in Canada
cchsflow variable name: SDCGRES}
}
\value{
value of percentage of time spent living in Canada in the Pct_time_derived variable
}
\description{
This function creates a derived variable that provides a percentage of the time a person's life was spent in Canada
}
