% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent-time-canada.R
\name{Pct_time_fun}
\alias{Pct_time_fun}
\title{Percent time in Canada}
\usage{
Pct_time_fun(DHHGAGE_cont, SDCGCBG, SDCGRES)
}
\arguments{
\item{DHHGAGE_cont}{continuous age variable.}

\item{SDCGCBG}{whether or not someone was born in Canada}

\item{SDCGRES}{how long someone has lived in Canada}
}
\value{
numeric value that is a fraction between 0 and 1 that represents percentage of a respondent's time in Canada
}
\description{
This function creates a derived variable that provides a percentage of the time a person's life was spent in 
 Canada.
}
