% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACDPAL.R
\name{RACDPAL_fun}
\alias{RACDPAL_fun}
\title{Participation and Activity Limitation}
\usage{
RACDPAL_fun(RAC_1, RAC_2A, RAC_2B, RAC_2C)
}
\arguments{
\item{RAC_1}{Has difficulty with activities due to disability}

\item{RAC_2A}{Reduction in activities at home due to disability}

\item{RAC_2B}{Reduction in activities at school or work due to disability}

\item{RAC_2C}{Reduction in other activities}
}
\value{
the CCHS derived variable RACDPAL with 3 categories:
 \enumerate{
  \item Sometimes
  \item Often
  \item Never
  }
}
\description{
This is a derived variable used in the CCHS (RACDPAL) to
 classify respondents according to the frequency with which they experience
 activity limitations due to disability.
}
\details{
This derived variable is generated in CCHS cycles 2003-2014. The
 2001 CCHS cycle, however, contains the same base variables used to derive
 this variable. To include respondents in the 2001 CCHS cycle, this custom
 function was created using the same derivation conditions used in later
 cycles.
}
\examples{
# Using RACDPAL_fun() to transform RACDPAL in 2001. 
# RACDPAL_fun() is specified in variable_details.csv along with the
# CCHS variables and cycles included.
 
# To transform RACDPAL, use rec_with_table() for each the 2001 cycle
# and specify RACDPAL, along with the various ADL variables.

library(cchsflow)

RACDPAL_2001 <- rec_with_table(
  cchs2001_p, c(
    "RAC_1", "RAC_2A", "RAC_2B", "RAC_2C", "RACDPAL"
  )
)

head(RACDPAL_2001)

# Note: In other CCHS cycles you only need to specify RACDPAL as the variable
# was included in those survey cycles.

}
