% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-with-table.R
\name{recode_columns}
\alias{recode_columns}
\title{recode_columns}
\usage{
recode_columns(
  data,
  variables_to_process,
  data_name,
  log,
  print_note,
  else_default
)
}
\arguments{
\item{data}{The source database}

\item{variables_to_process}{rows from variable details that are applicable
to this DB}

\item{data_name}{Name of the database being passed}

\item{log}{The option of printing log}

\item{print_note}{the option of printing the note columns}

\item{else_default}{default else value to use if no else is present}
}
\value{
Returns recoded and labeled data
}
\description{
Recodes columns from passed row and returns just table with those columns
and same rows as the data
}
