% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-with-table.R
\name{recode_variable_NA_formating}
\alias{recode_variable_NA_formating}
\title{Recode NA formatting}
\usage{
recode_variable_NA_formating(cell_value, var_type)
}
\arguments{
\item{cell_value}{The value inside the recTo column}

\item{var_type}{the toType of a variable}
}
\value{
an appropriately coded tagged NA
}
\description{
Recodes the NA depending on the var type
}
