% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{resp_condition_fun}
\alias{resp_condition_fun}
\title{resp_condition_fun}
\usage{
resp_condition_fun(age, CCC_091, CCC_031)
}
\arguments{
\item{age}{continuous age variable (DHHGAGE_cont for _p cycles, DHH_AGE for
_m cycles).}

\item{CCC_091}{variable indicating if respondent has COPD, emphysema, or
bronchitis}

\item{CCC_031}{variable indicating if respondent has asthma}
}
\value{
a categorical variable (resp_condition_der) with 3 levels:

 \enumerate{
 \item respondent is over the age of 35 and has a respiratory condition
 \item respondent is under the age of 35 and has a respiratory condition
 \item respondent does not have a respiratory condition
 }
}
\description{
This function is used to create a derived variable
 (resp_condition_der) that determines if a respondent has a respiratory
 condition. It uses the combined COPD, Emphysema, and Bronchitis variable
 (CCC_091) and Asthma (CCC_031) across all CCHS cycles.
}
\examples{
resp_condition_fun(40, 1, 1)  # 1 (over 35, has condition)
resp_condition_fun(25, 1, 1)  # 2 (under 35, has condition)
resp_condition_fun(40, 2, 2)  # 3 (no condition)

}
