% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{Resp_condition_fun2}
\alias{Resp_condition_fun2}
\title{Resp_condition_fun2}
\usage{
Resp_condition_fun2(DHHGAGE_cont, CCC_91E, CCC_91F, CCC_91A)
}
\arguments{
\item{DHHGAGE_cont}{continuous age variable.}

\item{CCC_91E}{variable indicating if respondent has emphysema}

\item{CCC_91F}{variable indicating if respondent has COPD}

\item{CCC_91A}{variable indicating if respondent has chronic bronchitis}
}
\value{
a categorical variable (Resp_condition_der) with 3 levels:

 \enumerate{
 \item respondent is over the age of 35 and has a respiratory condition
 \item respondent is under the age of 35 and has a respiratory condition
 \item respondent does not have a respiratory condition
 }
}
\description{
This is one of 3 functions used to create a derived variable (Resp_condition_der)
 that determines if a respondents has a respirtory condition. This function is for CCHS cycles
 (2005-2007) that use COPD & Emphysema as separate variables, as well as Bronchitis.
}
\examples{

# Using Resp_condition_fun2() to create pack-years values across CCHS cycles (2005-2007)
# Resp_condition_fun2() is specified in variableDetails.csv along with the CCHS variables and
# cycles included.

# To transform Resp_condition_der, use RecWTable() for each CCHS cycle and specify
# Resp_condition_der,
# along with the various respiratory variables. Then by using bind_rows(), you can combined
# Resp_condition_der across cycles.

library(cchsflow)

resp2005 <- rec_with_table(
  data_source = cchs2005, variable_details = variableDetails,
  dataset_name = "cchs2005", variables = c(
    "DHHGAGE_cont", "CCC_91E", "CCC_91F", "CCC_91A",
    "Resp_condition_der"
  )
)
head(resp2005)

resp2007_2008 <- rec_with_table(
  data_source = cchs2007_2008, variable_details = variableDetails,
  dataset_name = "cchs2007_2008", variables = c(
    "DHHGAGE_cont", "CCC_91E", "CCC_91F", "CCC_91A",
    "Resp_condition_der"
  )
)

tail(resp2007_2008)

combined_resp <- bind_rows(resp2005, resp2007_2008)

head(combined_resp)
tail(combined_resp)
}
\seealso{
\code{\link{Resp_condition_fun1}}, \code{\link{Resp_condition_fun3}}
}
