% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{resp_condition_fun3}
\alias{resp_condition_fun3}
\title{resp_condition_fun3}
\usage{
resp_condition_fun3(DHHGAGE_cont, CCC_091, CCC_91A)
}
\arguments{
\item{DHHGAGE_cont}{continuous age variable.}

\item{CCC_091}{variable indicating if respondent has either COPD or Emphysema}

\item{CCC_91A}{variable indicating if respondent has chronic bronchitis
cchsflow variable name: CCC_91A}
}
\value{
a categorical variable (resp_condition_der) with 3 levels:

 \enumerate{
 \item respondent is over the age of 35 and has a respiratory condition
 \item respondent is under the age of 35 and has a respiratory condition
 \item respondent does not have a respiratory condition
 }
}
\description{
This is one of 3 functions used to create a derived variable 
 (resp_condition_der) that determines if a respondents has a respirtory 
 condition. This function for CCHS cycles (2001-2003) that use COPD and 
 Emphysema as a combined variable, as well as Bronchitis
}
\seealso{
\code{\link{resp_condition_fun1}}, \code{\link{resp_condition_fun2}}
}
