% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-variable-functions.R
\name{RespCondition_fun1}
\alias{RespCondition_fun1}
\title{Respiratory condition variable}
\usage{
RespCondition_fun1(Age_cont, COPD_Emphys)
}
\arguments{
\item{Age_cont}{continuous age variable. 
cchsflow variable name: DHHGAGE_cont}

\item{COPD_Emphys}{variable indicating if subject has either COPD or Emphysema
cchsflow variable name: CCC_091}
}
\value{
a categorical variable with 3 levels:
1 - subject is over the age of 35 and has a respiratory condition
2 - subject is under the age of 35 and has a respiratory condition
3 - subject does not have a respiratory condition
}
\description{
This function creates a derived variable that determines if a subject has a respiratory condition
}
\details{
This function is for CCHS cycles that only use COPD and Emphysema as a combined variable
}
