% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory-condition.R
\name{RespCondition_fun2}
\alias{RespCondition_fun2}
\title{RespCondition_fun2}
\usage{
RespCondition_fun2(DHHGAGE_cont, CCC_91E, CCC_91F, CCC_91A)
}
\arguments{
\item{DHHGAGE_cont}{continuous age variable.}

\item{CCC_91E}{variable indicating if respondent has emphysema}

\item{CCC_91F}{variable indicating if respondent has COPD}

\item{CCC_91A}{variable indicating if respondent has chronic bronchitis}
}
\value{
a categorical variable (RespCondition_derived) with 3 levels:
 1 - respondent is over the age of 35 and has a respiratory condition
 2 - respondent is under the age of 35 and has a respiratory condition
 3 - respondent does not have a respiratory condition
}
\description{
This is one of 3 functions used to create a derived variable (RespCondition_derived) that determines if a respondents has a respirtory condition. 
 This function is for CCHS cycles (2005-2007) that use COPD & Emphysema as separate variables, as well as Bronchitis.
}
\seealso{
RespCondition_fun1, RespCondition_fun3
}
