% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-variable-functions.R
\name{RespCondition_fun3}
\alias{RespCondition_fun3}
\title{Respiratory condition variable}
\usage{
RespCondition_fun3(Age_cont, COPD_Emphys, Bronch)
}
\arguments{
\item{Age_cont}{continuous age variable. 
cchsflow variable name: DHHGAGE_cont}

\item{COPD_Emphys}{variable indicating if subject has either COPD or Emphysema
cchsflow variable name: CCC_091}

\item{Bronc}{variable indicating if subject has chronic bronchitis
cchsflow variable name: CCC_91A}
}
\value{
a categorical variable with 3 levels:
1 - subject is over the age of 35 and has a respiratory condition
2 - subject is under the age of 35 and has a respiratory condition
3 - subject does not have a respiratory condition
}
\description{
This function for CCHS cycles that use COPD and Emphysema as a combined variable, as well as Bronchitis
}
