% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{sedentary_activity2_fun}
\alias{sedentary_activity2_fun}
\title{Total hours/week - sedentary activity excluding reading (age 20+)}
\usage{
sedentary_activity2_fun(DHHGAGE_cont, SAC_1, SAC_2_A, SAC_3)
}
\arguments{
\item{DHHGAGE_cont}{continuous age}

\item{SAC_1}{number of hours/week - spent on computer}

\item{SAC_2_A}{number of hours/week - playing video games (age 20+)}

\item{SAC_3}{number of hours/week - watching TV or videos}
}
\value{
value for total hours/week doing sedentary activities (excluding 
reading)
}
\description{
This function creates a derived variable (sedentary_activity2)
 that calculates the total number of hours in a week spent doing sedentary 
 activities such as using a computer, using the internet, playing video 
 games, and watching television or videos. For all activities, the 
 time spent at school or work is excluded. Time spent in reading is not 
 included. This is limited to ages 20+.
}
