% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{sedentary_activity_fun}
\alias{sedentary_activity_fun}
\title{Total hours/week - sedentary activity (age 20+)}
\usage{
sedentary_activity_fun(DHHGAGE_cont, SAC_1, SAC_2_A, SAC_3, SAC_4)
}
\arguments{
\item{DHHGAGE_cont}{continuous age}

\item{SAC_1}{number of hours/week - spent on computer}

\item{SAC_2_A}{number of hours/week - playing video games (age 20+)}

\item{SAC_3}{number of hours/week - watching TV or videos}

\item{SAC_4}{number of hours/week - spent reading}
}
\value{
value for total hours/week doing sedentary activities
}
\description{
This function creates a derived variable (sedentary_activity)
 that calculates the total number of hours in a week spent doing sedentary 
 activities such as using a computer, using the internet, playing video 
 games, watching television or videos, and reading. For all activities, the 
 time spent at school or work is excluded. This is limited to ages 20+.
}
\note{
This is limited to ages 20+ for CCHS 2009-2016.
}
\examples{
# Using sedentary_activity_fun() to create values for total hours per week 
# spent doing sedentary activities across CCHS cycles.
# sedentary_activity_fun() is specified in variable_details.csv along with 
# the CCHS variables and cycles included.

# To transform sedentary_activity across cycles, use rec_with_table() for each
# CCHS cycle and specify sedentary_activity, along with the required base
# variables.Then by using merge_rec_data(), you can combine 
# sedentary_activity across cycles.

library(cchsflow)

sed_act2009_2010 <- rec_with_table(
  cchs2009_2010_p, c(
    "SAC_1", "SAC_2_A","SAC_3","SAC_4","DHHGAGE_cont","sedentary_activity"
  )
)

head(sed_act2009_2010)

sed_act2015_2016 <- rec_with_table(
  cchs2015_2016_p, c(
    "SAC_1", "SAC_2D_A","SAC_2D_B","SAC_2D","DHHGAGE_cont","SAC_2_A",
    "SAC_3","SAC_4","sedentary_activity"
  )
)
tail(sed_act2015_2016)

combined_sed_act <- suppressWarnings(merge_rec_data(sed_act2009_2010,
 sed_act2015_2016))

head(combined_sed_act)
tail(combined_sed_act)
}
