% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-utils.R
\name{set_data_labels}
\alias{set_data_labels}
\title{Set Data Labels}
\usage{
set_data_labels(data_to_label, variable_details, variables_sheet = NULL)
}
\arguments{
\item{data_to_label}{newly transformed dataset}

\item{variable_details}{variableDetails.csv}

\item{variables_sheet}{variables.csv}
}
\description{
sets labels for transformed datasets
}
