% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{SMKG203_fun}
\alias{SMKG203_fun}
\title{Age started to smoke daily - daily smoker (from combined variable)}
\usage{
SMKG203_fun(SMK_005, SMKG040)
}
\arguments{
\item{SMK_005}{type of smoker presently}

\item{SMKG040}{age started to smoke daily - daily/former daily smoker}
}
\value{
value for continuous age started to smoke daily for daily smokers
}
\description{
This function creates a continuous derived variable
(SMKG203_cont) for age started to smoke daily for daily smokers.
}
