% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{SMKG207_fun}
\alias{SMKG207_fun}
\title{Age started to smoke daily - former daily smoker (from combined variable)}
\usage{
SMKG207_fun(SMK_030, SMKG040)
}
\arguments{
\item{SMK_030}{smoked daily - lifetime (occasional/former smoker)}

\item{SMKG040}{age started to smoke daily - daily/former daily smoker}
}
\value{
value for continuous age started to smoke daily for former daily smokers
}
\description{
This function creates a continuous derived variable
(SMKG207_cont) for age started to smoke daily for former daily smokers.
}
