% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variableDetails}
\alias{variableDetails}
\title{variableDetails.csv}
\value{
\item{variableDetails}{a data frame}
}
\description{
This dataset provides details on how variables are specified in cchsflow.
}
\details{
See \href{https://big-life-lab.github.io/cchsflow/articles/variablesDetails.html}{here} for more
details about how the worksheet is structured
}
\examples{
data(variableDetails)
str(variableDetails)
}
\keyword{datasets}
