% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary-activity.R
\name{weekly_screen_time_fun}
\alias{weekly_screen_time_fun}
\title{Weekly leisure screen time}
\usage{
weekly_screen_time_fun(SBE_010)
}
\arguments{
\item{SBE_010}{Daily screen time (hours/day) outside of school and work day for
CCHS 2017-2018}
}
\value{
value for screen time (hours/week) outside of school and work day
}
\description{
This function creates a derived variable (weekly_screen)
 that calculates the total hours in a week spent using a screen outside of 
 school and/or work.
}
\examples{
# Using weekly_screen_time_fun() to create values for weekly screen time 
# across CCHS cycles.
# weekly_screen_time_fun() is specified in variable_details.csv along with 
# the CCHS variables and cycles included.

# To transform weekly_screen across cycles, use rec_with_table() for each
# CCHS cycle and specify weekly_screen, along with the daily screen time
# variable.Then by using merge_rec_data(), you can combine 
# weekly_screen across cycles.

library(cchsflow)

weekly_st2017_2018 <- rec_with_table(
  cchs2017_2018_p, c(
    "SBE_010", "weekly_screen"
  )
)

head(weekly_st2017_2018)
tail(weekly_st2017_2018)
}
